Crate rusoto_ssm

AWS Systems Manager
AWS Systems Manager is a collection of capabilities that helps you automate management tasks such as collecting system inventory, applying operating system (OS) patches, automating the creation of Amazon Machine Images (AMIs), and configuring operating systems (OSs) and applications at scale. Systems Manager lets you remotely and securely manage the configuration of your managed instances. A managed instance is any Amazon EC2 instance or on-premises machine in your hybrid environment that has been configured for Systems Manager.

This reference is intended to be used with the AWS Systems Manager User Guide.

To get started, verify prerequisites and configure managed instances. For more information, see Setting Up AWS Systems Manager in the AWS Systems Manager User Guide.

For information about other API actions you can perform on Amazon EC2 instances, see the Amazon EC2 API Reference. For information about how to use a Query API, see Making API Requests.

If you're using the service, you're probably looking for SsmClient and Ssm.

Structs

Activation
An activation registers one or more on-premises servers or virtual machines (VMs) with AWS so that you can configure those servers or VMs using Run Command. A server or VM that has been registered with AWS is called a managed instance.

AddTagsToResourceRequest

AddTagsToResourceResult

Association
Describes an association of a Systems Manager document and an instance.

AssociationDescription
Describes the parameters for a document.

AssociationExecution
Includes information about the specified association.

AssociationExecutionFilter
Filters used in the request.

AssociationExecutionTarget
Includes information about the specified association execution.

AssociationExecutionTargetsFilter
Filters for the association execution.

AssociationFilter
Describes a filter.

AssociationOverview
Information about the association.

AssociationStatus
Describes an association status.

AssociationVersionInfo
Information about the association version.

AttachmentContent
A structure that includes attributes that describe a document attachment.

AttachmentInformation
An attribute of an attachment, such as the attachment name.

AttachmentsSource
A key and value pair that identifies the location of an attachment to a document.

AutomationExecution
Detailed information about the current state of an individual Automation execution.

AutomationExecutionFilter
A filter used to match specific automation executions. This is used to limit the scope of Automation execution information returned.

AutomationExecutionMetadata
Details about a specific Automation execution.

CancelCommandRequest

CancelCommandResult
Whether or not the command was successfully canceled. There is no guarantee that a request can be canceled.

CancelMaintenanceWindowExecutionRequest

CancelMaintenanceWindowExecutionResult

CloudWatchOutputConfig
Configuration options for sending command output to CloudWatch Logs.

Command
Describes a command request.

CommandFilter
Describes a command filter.

CommandInvocation
An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user runs SendCommand against three instances, then a command invocation is created for each requested instance ID. A command invocation returns status and detail information about a command you ran.

CommandPlugin
Describes plugin details.

ComplianceExecutionSummary
A summary of the call execution that includes an execution ID, the type of execution (for example, Command), and the date/time of the execution using a datetime object that is saved in the following format: yyyy-MM-dd'T'HH:mm:ss'Z'.

ComplianceItem
Information about the compliance as defined by the resource type. For example, for a patch resource type, Items includes information about the PatchSeverity, Classification, etc.

ComplianceItemEntry
Information about a compliance item.

ComplianceStringFilter
One or more filters. Use a filter to return a more specific list of results.

ComplianceSummaryItem
A summary of compliance information by compliance type.

CompliantSummary
A summary of resources that are compliant. The summary is organized according to the resource count for each compliance type.

CreateActivationRequest

CreateActivationResult

CreateAssociationBatchRequest

CreateAssociationBatchRequestEntry
Describes the association of a Systems Manager SSM document and an instance.

CreateAssociationBatchResult

CreateAssociationRequest

CreateAssociationResult

CreateDocumentRequest

CreateDocumentResult

CreateMaintenanceWindowRequest

CreateMaintenanceWindowResult

CreateOpsItemRequest

CreateOpsItemResponse

CreatePatchBaselineRequest

CreatePatchBaselineResult

CreateResourceDataSyncRequest

CreateResourceDataSyncResult

DeleteActivationRequest

DeleteActivationResult

DeleteAssociationRequest

DeleteAssociationResult

DeleteDocumentRequest

DeleteDocumentResult

DeleteInventoryRequest

DeleteInventoryResult

DeleteMaintenanceWindowRequest

DeleteMaintenanceWindowResult

DeleteParameterRequest

DeleteParameterResult

DeleteParametersRequest

DeleteParametersResult

DeletePatchBaselineRequest

DeletePatchBaselineResult

DeleteResourceDataSyncRequest

DeleteResourceDataSyncResult

DeregisterManagedInstanceRequest

DeregisterManagedInstanceResult

DeregisterPatchBaselineForPatchGroupRequest

DeregisterPatchBaselineForPatchGroupResult

DeregisterTargetFromMaintenanceWindowRequest

DeregisterTargetFromMaintenanceWindowResult

DeregisterTaskFromMaintenanceWindowRequest

DeregisterTaskFromMaintenanceWindowResult

DescribeActivationsFilter
Filter for the DescribeActivation API.

DescribeActivationsRequest

DescribeActivationsResult

DescribeAssociationExecutionTargetsRequest

DescribeAssociationExecutionTargetsResult

DescribeAssociationExecutionsRequest

DescribeAssociationExecutionsResult

DescribeAssociationRequest

DescribeAssociationResult

DescribeAutomationExecutionsRequest

DescribeAutomationExecutionsResult

DescribeAutomationStepExecutionsRequest

DescribeAutomationStepExecutionsResult

DescribeAvailablePatchesRequest

DescribeAvailablePatchesResult

DescribeDocumentPermissionRequest

DescribeDocumentPermissionResponse

DescribeDocumentRequest

DescribeDocumentResult

DescribeEffectiveInstanceAssociationsRequest

DescribeEffectiveInstanceAssociationsResult

DescribeEffectivePatchesForPatchBaselineRequest

DescribeEffectivePatchesForPatchBaselineResult

DescribeInstanceAssociationsStatusRequest

DescribeInstanceAssociationsStatusResult

DescribeInstanceInformationRequest

DescribeInstanceInformationResult

DescribeInstancePatchStatesForPatchGroupRequest

DescribeInstancePatchStatesForPatchGroupResult

DescribeInstancePatchStatesRequest

DescribeInstancePatchStatesResult

DescribeInstancePatchesRequest

DescribeInstancePatchesResult

DescribeInventoryDeletionsRequest

DescribeInventoryDeletionsResult

DescribeMaintenanceWindowExecutionTaskInvocationsRequest

DescribeMaintenanceWindowExecutionTaskInvocationsResult

DescribeMaintenanceWindowExecutionTasksRequest

DescribeMaintenanceWindowExecutionTasksResult

DescribeMaintenanceWindowExecutionsRequest

DescribeMaintenanceWindowExecutionsResult

DescribeMaintenanceWindowScheduleRequest

DescribeMaintenanceWindowScheduleResult

DescribeMaintenanceWindowTargetsRequest

DescribeMaintenanceWindowTargetsResult

DescribeMaintenanceWindowTasksRequest

DescribeMaintenanceWindowTasksResult

DescribeMaintenanceWindowsForTargetRequest

DescribeMaintenanceWindowsForTargetResult

DescribeMaintenanceWindowsRequest

DescribeMaintenanceWindowsResult

DescribeOpsItemsRequest

DescribeOpsItemsResponse

DescribeParametersRequest

DescribeParametersResult

DescribePatchBaselinesRequest

DescribePatchBaselinesResult

DescribePatchGroupStateRequest

DescribePatchGroupStateResult

DescribePatchGroupsRequest

DescribePatchGroupsResult

DescribePatchPropertiesRequest

DescribePatchPropertiesResult

DescribeSessionsRequest

DescribeSessionsResponse

DocumentDefaultVersionDescription
A default version of a document.

DocumentDescription
Describes a Systems Manager document.

DocumentFilter
Describes a filter.

DocumentIdentifier
Describes the name of a Systems Manager document.

DocumentKeyValuesFilter
One or more filters. Use a filter to return a more specific list of documents.

For keys, you can specify one or more tags that have been applied to a document.

Other valid values include Owner, Name, PlatformTypes, and DocumentType.

Note that only one Owner can be specified in a request. For example: Key=Owner,Values=Self.

If you use Name as a key, you can use a name prefix to return a list of documents. For example, in the AWS CLI, to return a list of all documents that begin with Te, run the following command:

aws ssm list-documents --filters Key=Name,Values=Te

If you specify more than two keys, only documents that are identified by all the tags are returned in the results. If you specify more than two values for a key, documents that are identified by any of the values are returned in the results.

To specify a custom key and value pair, use the format Key=tag:[tagName],Values=[valueName].

For example, if you created a Key called region and are using the AWS CLI to call the list-documents command:

aws ssm list-documents --filters Key=tag:region,Values=east,west Key=Owner,Values=Self

DocumentParameter
Parameters specified in a System Manager document that run on the server when the command is run.

DocumentVersionInfo
Version information about the document.

EffectivePatch
The EffectivePatch structure defines metadata about a patch along with the approval state of the patch in a particular patch baseline. The approval state includes information about whether the patch is currently approved, due to be approved by a rule, explicitly approved, or explicitly rejected and the date the patch was or will be approved.

FailedCreateAssociation
Describes a failed association.

FailureDetails
Information about an Automation failure.

GetAutomationExecutionRequest

GetAutomationExecutionResult

GetCommandInvocationRequest

GetCommandInvocationResult

GetConnectionStatusRequest

GetConnectionStatusResponse

GetDefaultPatchBaselineRequest

GetDefaultPatchBaselineResult

GetDeployablePatchSnapshotForInstanceRequest

GetDeployablePatchSnapshotForInstanceResult

GetDocumentRequest

GetDocumentResult

GetInventoryRequest

GetInventoryResult

GetInventorySchemaRequest

GetInventorySchemaResult

GetMaintenanceWindowExecutionRequest

GetMaintenanceWindowExecutionResult

GetMaintenanceWindowExecutionTaskInvocationRequest

GetMaintenanceWindowExecutionTaskInvocationResult

GetMaintenanceWindowExecutionTaskRequest

GetMaintenanceWindowExecutionTaskResult

GetMaintenanceWindowRequest

GetMaintenanceWindowResult

GetMaintenanceWindowTaskRequest

GetMaintenanceWindowTaskResult

GetOpsItemRequest

GetOpsItemResponse

GetOpsSummaryRequest

GetOpsSummaryResult

GetParameterHistoryRequest

GetParameterHistoryResult

GetParameterRequest

GetParameterResult

GetParametersByPathRequest

GetParametersByPathResult

GetParametersRequest

GetParametersResult

GetPatchBaselineForPatchGroupRequest

GetPatchBaselineForPatchGroupResult

GetPatchBaselineRequest

GetPatchBaselineResult

GetServiceSettingRequest
The request body of the GetServiceSetting API action.

GetServiceSettingResult
The query result body of the GetServiceSetting API action.

InstanceAggregatedAssociationOverview
Status information about the aggregated associations.

InstanceAssociation
One or more association documents on the instance.

InstanceAssociationOutputLocation
An Amazon S3 bucket where you want to store the results of this request.

InstanceAssociationOutputUrl
The URL of Amazon S3 bucket where you want to store the results of this request.

InstanceAssociationStatusInfo
Status information about the instance association.

InstanceInformation
Describes a filter for a specific list of instances.

InstanceInformationFilter
Describes a filter for a specific list of instances. You can filter instances information by using tags. You specify tags by using a key-value mapping.

Use this action instead of the DescribeInstanceInformationRequest$InstanceInformationFilterList method. The InstanceInformationFilterList method is a legacy method and does not support tags.

InstanceInformationStringFilter
The filters to describe or get information about your managed instances.

InstancePatchState
Defines the high-level patch compliance state for a managed instance, providing information about the number of installed, missing, not applicable, and failed patches along with metadata about the operation when this information was gathered for the instance.

InstancePatchStateFilter
Defines a filter used in DescribeInstancePatchStatesForPatchGroup used to scope down the information returned by the API.

InventoryAggregator
Specifies the inventory type and attribute for the aggregation execution.

InventoryDeletionStatusItem
Status information returned by the DeleteInventory action.

InventoryDeletionSummary
Information about the delete operation.

InventoryDeletionSummaryItem
Either a count, remaining count, or a version number in a delete inventory summary.

InventoryFilter
One or more filters. Use a filter to return a more specific list of results.

InventoryGroup
A user-defined set of one or more filters on which to aggregate inventory data. Groups return a count of resources that match and don't match the specified criteria.

InventoryItem
Information collected from managed instances based on your inventory policy document

InventoryItemAttribute
Attributes are the entries within the inventory item content. It contains name and value.

InventoryItemSchema
The inventory item schema definition. Users can use this to compose inventory query filters.

InventoryResultEntity
Inventory query results.

InventoryResultItem
The inventory result item.

LabelParameterVersionRequest

LabelParameterVersionResult

ListAssociationVersionsRequest

ListAssociationVersionsResult

ListAssociationsRequest

ListAssociationsResult

ListCommandInvocationsRequest

ListCommandInvocationsResult

ListCommandsRequest

ListCommandsResult

ListComplianceItemsRequest

ListComplianceItemsResult

ListComplianceSummariesRequest

ListComplianceSummariesResult

ListDocumentVersionsRequest

ListDocumentVersionsResult

ListDocumentsRequest

ListDocumentsResult

ListInventoryEntriesRequest

ListInventoryEntriesResult

ListResourceComplianceSummariesRequest

ListResourceComplianceSummariesResult

ListResourceDataSyncRequest

ListResourceDataSyncResult

ListTagsForResourceRequest

ListTagsForResourceResult

LoggingInfo

Information about an Amazon S3 bucket to write instance-level logs to.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

MaintenanceWindowAutomationParameters
The parameters for an AUTOMATION task type.

MaintenanceWindowExecution
Describes the information about an execution of a maintenance window.

MaintenanceWindowExecutionTaskIdentity
Information about a task execution performed as part of a maintenance window execution.

MaintenanceWindowExecutionTaskInvocationIdentity
Describes the information about a task invocation for a particular target as part of a task execution performed as part of a maintenance window execution.

MaintenanceWindowFilter
Filter used in the request. Supported filter keys are Name and Enabled.

MaintenanceWindowIdentity
Information about the maintenance window.

MaintenanceWindowIdentityForTarget
The maintenance window to which the specified target belongs.

MaintenanceWindowLambdaParameters

The parameters for a LAMBDA task type.

For information about specifying and updating task parameters, see RegisterTaskWithMaintenanceWindow and UpdateMaintenanceWindowTask.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

For Lambda tasks, Systems Manager ignores any values specified for TaskParameters and LoggingInfo.

MaintenanceWindowRunCommandParameters

The parameters for a RUN_COMMAND task type.

For information about specifying and updating task parameters, see RegisterTaskWithMaintenanceWindow and UpdateMaintenanceWindowTask.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

For Run Command tasks, Systems Manager uses specified values for TaskParameters and LoggingInfo only if no values are specified for TaskInvocationParameters.

MaintenanceWindowStepFunctionsParameters

The parameters for a STEP_FUNCTION task.

For information about specifying and updating task parameters, see RegisterTaskWithMaintenanceWindow and UpdateMaintenanceWindowTask.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

For Step Functions tasks, Systems Manager ignores any values specified for TaskParameters and LoggingInfo.

MaintenanceWindowTarget
The target registered with the maintenance window.

MaintenanceWindowTask
Information about a task defined for a maintenance window.

MaintenanceWindowTaskInvocationParameters
The parameters for task execution.

MaintenanceWindowTaskParameterValueExpression
Defines the values for a task parameter.

ModifyDocumentPermissionRequest

ModifyDocumentPermissionResponse

NonCompliantSummary
A summary of resources that are not compliant. The summary is organized according to resource type.

NotificationConfig
Configurations for sending notifications.

OpsAggregator
One or more aggregators for viewing counts of OpsItems using different dimensions such as Source, CreatedTime, or Source and CreatedTime, to name a few.

OpsEntity
The result of the query.

OpsEntityItem
The OpsItem summaries result item.

OpsFilter
A filter for viewing OpsItem summaries.

OpsItem
Operations engineers and IT professionals use the Systems Manager OpsItems capability to view, investigate, and remediate operational issues impacting the performance and health of their AWS resources. For more information, see AWS Systems Manager OpsItems in the AWS Systems Manager User Guide.

OpsItemDataValue
An object that defines the value of the key and its type in the OperationalData map.

OpsItemFilter
Describes an OpsCenter filter.

OpsItemNotification
A notification about the OpsItem.

OpsItemSummary
A count of OpsItems.

OutputSource
Information about the source where the association execution details are stored.

Parameter
An Amazon EC2 Systems Manager parameter in Parameter Store.

ParameterHistory
Information about parameter usage.

ParameterInlinePolicy
One or more policies assigned to a parameter.

ParameterMetadata
Metadata includes information like the ARN of the last user and the date/time the parameter was last used.

ParameterStringFilter

One or more filters. Use a filter to return a more specific list of results.

The Name and Tier filter keys can't be used with the GetParametersByPath API action. Also, the Label filter key can't be used with the DescribeParameters API action.

ParametersFilter
This data type is deprecated. Instead, use ParameterStringFilter.

Patch
Represents metadata about a patch.

PatchBaselineIdentity
Defines the basic information about a patch baseline.

PatchComplianceData
Information about the state of a patch on a particular instance as it relates to the patch baseline used to patch the instance.

PatchFilter
Defines which patches should be included in a patch baseline.

A patch filter consists of a key and a set of values. The filter key is a patch property. For example, the available filter keys for WINDOWS are PATCH_SET, PRODUCT, PRODUCT_FAMILY, CLASSIFICATION, and MSRC_SEVERITY. The filter values define a matching criterion for the patch property indicated by the key. For example, if the filter key is PRODUCT and the filter values are ["Office 2013", "Office 2016"], then the filter accepts all patches where product name is either "Office 2013" or "Office 2016". The filter values can be exact values for the patch property given as a key, or a wildcard (*), which matches all values.

You can view lists of valid values for the patch properties by running the DescribePatchProperties command. For information about which patch properties can be used with each major operating system, see DescribePatchProperties.

PatchFilterGroup
A set of patch filters, typically used for approval rules.

PatchGroupPatchBaselineMapping
The mapping between a patch group and the patch baseline the patch group is registered with.

PatchOrchestratorFilter
Defines a filter used in Patch Manager APIs.

PatchRule
Defines an approval rule for a patch baseline.

PatchRuleGroup
A set of rules defining the approval rules for a patch baseline.

PatchSource
Information about the patches to use to update the instances, including target operating systems and source repository. Applies to Linux instances only.

PatchStatus
Information about the approval status of a patch.

ProgressCounters
An aggregate of step execution statuses displayed in the AWS Console for a multi-Region and multi-account Automation execution.

PutComplianceItemsRequest

PutComplianceItemsResult

PutInventoryRequest

PutInventoryResult

PutParameterRequest

PutParameterResult

RegisterDefaultPatchBaselineRequest

RegisterDefaultPatchBaselineResult

RegisterPatchBaselineForPatchGroupRequest

RegisterPatchBaselineForPatchGroupResult

RegisterTargetWithMaintenanceWindowRequest

RegisterTargetWithMaintenanceWindowResult

RegisterTaskWithMaintenanceWindowRequest

RegisterTaskWithMaintenanceWindowResult

RelatedOpsItem
An OpsItems that shares something in common with the current OpsItems. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

RemoveTagsFromResourceRequest

RemoveTagsFromResourceResult

ResetServiceSettingRequest
The request body of the ResetServiceSetting API action.

ResetServiceSettingResult
The result body of the ResetServiceSetting API action.

ResolvedTargets
Information about targets that resolved during the Automation execution.

ResourceComplianceSummaryItem
Compliance summary information for a specific resource.

ResourceDataSyncItem
Information about a Resource Data Sync configuration, including its current status and last successful sync.

ResourceDataSyncS3Destination
Information about the target Amazon S3 bucket for the Resource Data Sync.

ResultAttribute
The inventory item result attribute.

ResumeSessionRequest

ResumeSessionResponse

S3OutputLocation
An Amazon S3 bucket where you want to store the results of this request.

S3OutputUrl
A URL for the Amazon S3 bucket where you want to store the results of this request.

ScheduledWindowExecution
Information about a scheduled execution for a maintenance window.

SendAutomationSignalRequest

SendAutomationSignalResult

SendCommandRequest

SendCommandResult

ServiceSetting
The service setting data structure.

ServiceSetting is an account-level setting for an AWS service. This setting defines how a user interacts with or uses a service or a feature of a service. For example, if an AWS service charges money to the account based on feature or service usage, then the AWS service team might create a default setting of "false". This means the user can't use this feature unless they change the setting to "true" and intentionally opt in for a paid feature.

Services map a SettingId object to a setting value. AWS services teams define the default value for a SettingId. You can't create a new SettingId, but you can overwrite the default value if you have the ssm:UpdateServiceSetting permission for the setting. Use the UpdateServiceSetting API action to change the default setting. Or, use the ResetServiceSetting to change the value back to the original value defined by the AWS service team.

Session
Information about a Session Manager connection to an instance.

SessionFilter
Describes a filter for Session Manager information.

SessionManagerOutputUrl
Reserved for future use.

SeveritySummary
The number of managed instances found for each patch severity level defined in the request filter.

SsmClient
A client for the Amazon SSM API.

StartAssociationsOnceRequest

StartAssociationsOnceResult

StartAutomationExecutionRequest

StartAutomationExecutionResult

StartSessionRequest

StartSessionResponse

StepExecution
Detailed information about an the execution state of an Automation step.

StepExecutionFilter
A filter to limit the amount of step execution information returned by the call.

StopAutomationExecutionRequest

StopAutomationExecutionResult

Tag
Metadata that you assign to your AWS resources. Tags enable you to categorize your resources in different ways, for example, by purpose, owner, or environment. In Systems Manager, you can apply tags to documents, managed instances, maintenance windows, Parameter Store parameters, and patch baselines.

Target

An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call.

TargetLocation
The combination of AWS Regions and accounts targeted by the current Automation execution.

TerminateSessionRequest

TerminateSessionResponse

UpdateAssociationRequest

UpdateAssociationResult

UpdateAssociationStatusRequest

UpdateAssociationStatusResult

UpdateDocumentDefaultVersionRequest

UpdateDocumentDefaultVersionResult

UpdateDocumentRequest

UpdateDocumentResult

UpdateMaintenanceWindowRequest

UpdateMaintenanceWindowResult

UpdateMaintenanceWindowTargetRequest

UpdateMaintenanceWindowTargetResult

UpdateMaintenanceWindowTaskRequest

UpdateMaintenanceWindowTaskResult

UpdateManagedInstanceRoleRequest

UpdateManagedInstanceRoleResult

UpdateOpsItemRequest

UpdateOpsItemResponse

UpdatePatchBaselineRequest

UpdatePatchBaselineResult

UpdateServiceSettingRequest
The request body of the UpdateServiceSetting API action.

UpdateServiceSettingResult
The result body of the UpdateServiceSetting API action.

Enums

AddTagsToResourceError
Errors returned by AddTagsToResource

CancelCommandError
Errors returned by CancelCommand

CancelMaintenanceWindowExecutionError
Errors returned by CancelMaintenanceWindowExecution

CreateActivationError
Errors returned by CreateActivation

CreateAssociationBatchError
Errors returned by CreateAssociationBatch

CreateAssociationError
Errors returned by CreateAssociation

CreateDocumentError
Errors returned by CreateDocument

CreateMaintenanceWindowError
Errors returned by CreateMaintenanceWindow

CreateOpsItemError
Errors returned by CreateOpsItem

CreatePatchBaselineError
Errors returned by CreatePatchBaseline

CreateResourceDataSyncError
Errors returned by CreateResourceDataSync

DeleteActivationError
Errors returned by DeleteActivation

DeleteAssociationError
Errors returned by DeleteAssociation

DeleteDocumentError
Errors returned by DeleteDocument

DeleteInventoryError
Errors returned by DeleteInventory

DeleteMaintenanceWindowError
Errors returned by DeleteMaintenanceWindow

DeleteParameterError
Errors returned by DeleteParameter

DeleteParametersError
Errors returned by DeleteParameters

DeletePatchBaselineError
Errors returned by DeletePatchBaseline

DeleteResourceDataSyncError
Errors returned by DeleteResourceDataSync

DeregisterManagedInstanceError
Errors returned by DeregisterManagedInstance

DeregisterPatchBaselineForPatchGroupError
Errors returned by DeregisterPatchBaselineForPatchGroup

DeregisterTargetFromMaintenanceWindowError
Errors returned by DeregisterTargetFromMaintenanceWindow

DeregisterTaskFromMaintenanceWindowError
Errors returned by DeregisterTaskFromMaintenanceWindow

DescribeActivationsError
Errors returned by DescribeActivations

DescribeAssociationError
Errors returned by DescribeAssociation

DescribeAssociationExecutionTargetsError
Errors returned by DescribeAssociationExecutionTargets

DescribeAssociationExecutionsError
Errors returned by DescribeAssociationExecutions

DescribeAutomationExecutionsError
Errors returned by DescribeAutomationExecutions

DescribeAutomationStepExecutionsError
Errors returned by DescribeAutomationStepExecutions

DescribeAvailablePatchesError
Errors returned by DescribeAvailablePatches

DescribeDocumentError
Errors returned by DescribeDocument

DescribeDocumentPermissionError
Errors returned by DescribeDocumentPermission

DescribeEffectiveInstanceAssociationsError
Errors returned by DescribeEffectiveInstanceAssociations

DescribeEffectivePatchesForPatchBaselineError
Errors returned by DescribeEffectivePatchesForPatchBaseline

DescribeInstanceAssociationsStatusError
Errors returned by DescribeInstanceAssociationsStatus

DescribeInstanceInformationError
Errors returned by DescribeInstanceInformation

DescribeInstancePatchStatesError
Errors returned by DescribeInstancePatchStates

DescribeInstancePatchStatesForPatchGroupError
Errors returned by DescribeInstancePatchStatesForPatchGroup

DescribeInstancePatchesError
Errors returned by DescribeInstancePatches

DescribeInventoryDeletionsError
Errors returned by DescribeInventoryDeletions

DescribeMaintenanceWindowExecutionTaskInvocationsError
Errors returned by DescribeMaintenanceWindowExecutionTaskInvocations

DescribeMaintenanceWindowExecutionTasksError
Errors returned by DescribeMaintenanceWindowExecutionTasks

DescribeMaintenanceWindowExecutionsError
Errors returned by DescribeMaintenanceWindowExecutions

DescribeMaintenanceWindowScheduleError
Errors returned by DescribeMaintenanceWindowSchedule

DescribeMaintenanceWindowTargetsError
Errors returned by DescribeMaintenanceWindowTargets

DescribeMaintenanceWindowTasksError
Errors returned by DescribeMaintenanceWindowTasks

DescribeMaintenanceWindowsError
Errors returned by DescribeMaintenanceWindows

DescribeMaintenanceWindowsForTargetError
Errors returned by DescribeMaintenanceWindowsForTarget

DescribeOpsItemsError
Errors returned by DescribeOpsItems

DescribeParametersError
Errors returned by DescribeParameters

DescribePatchBaselinesError
Errors returned by DescribePatchBaselines

DescribePatchGroupStateError
Errors returned by DescribePatchGroupState

DescribePatchGroupsError
Errors returned by DescribePatchGroups

DescribePatchPropertiesError
Errors returned by DescribePatchProperties

DescribeSessionsError
Errors returned by DescribeSessions

GetAutomationExecutionError
Errors returned by GetAutomationExecution

GetCommandInvocationError
Errors returned by GetCommandInvocation

GetConnectionStatusError
Errors returned by GetConnectionStatus

GetDefaultPatchBaselineError
Errors returned by GetDefaultPatchBaseline

GetDeployablePatchSnapshotForInstanceError
Errors returned by GetDeployablePatchSnapshotForInstance

GetDocumentError
Errors returned by GetDocument

GetInventoryError
Errors returned by GetInventory

GetInventorySchemaError
Errors returned by GetInventorySchema

GetMaintenanceWindowError
Errors returned by GetMaintenanceWindow

GetMaintenanceWindowExecutionError
Errors returned by GetMaintenanceWindowExecution

GetMaintenanceWindowExecutionTaskError
Errors returned by GetMaintenanceWindowExecutionTask

GetMaintenanceWindowExecutionTaskInvocationError
Errors returned by GetMaintenanceWindowExecutionTaskInvocation

GetMaintenanceWindowTaskError
Errors returned by GetMaintenanceWindowTask

GetOpsItemError
Errors returned by GetOpsItem

GetOpsSummaryError
Errors returned by GetOpsSummary

GetParameterError
Errors returned by GetParameter

GetParameterHistoryError
Errors returned by GetParameterHistory

GetParametersByPathError
Errors returned by GetParametersByPath

GetParametersError
Errors returned by GetParameters

GetPatchBaselineError
Errors returned by GetPatchBaseline

GetPatchBaselineForPatchGroupError
Errors returned by GetPatchBaselineForPatchGroup

GetServiceSettingError
Errors returned by GetServiceSetting

LabelParameterVersionError
Errors returned by LabelParameterVersion

ListAssociationVersionsError
Errors returned by ListAssociationVersions

ListAssociationsError
Errors returned by ListAssociations

ListCommandInvocationsError
Errors returned by ListCommandInvocations

ListCommandsError
Errors returned by ListCommands

ListComplianceItemsError
Errors returned by ListComplianceItems

ListComplianceSummariesError
Errors returned by ListComplianceSummaries

ListDocumentVersionsError
Errors returned by ListDocumentVersions

ListDocumentsError
Errors returned by ListDocuments

ListInventoryEntriesError
Errors returned by ListInventoryEntries

ListResourceComplianceSummariesError
Errors returned by ListResourceComplianceSummaries

ListResourceDataSyncError
Errors returned by ListResourceDataSync

ListTagsForResourceError
Errors returned by ListTagsForResource

ModifyDocumentPermissionError
Errors returned by ModifyDocumentPermission

PutComplianceItemsError
Errors returned by PutComplianceItems

PutInventoryError
Errors returned by PutInventory

PutParameterError
Errors returned by PutParameter

RegisterDefaultPatchBaselineError
Errors returned by RegisterDefaultPatchBaseline

RegisterPatchBaselineForPatchGroupError
Errors returned by RegisterPatchBaselineForPatchGroup

RegisterTargetWithMaintenanceWindowError
Errors returned by RegisterTargetWithMaintenanceWindow

RegisterTaskWithMaintenanceWindowError
Errors returned by RegisterTaskWithMaintenanceWindow

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

ResetServiceSettingError
Errors returned by ResetServiceSetting

ResumeSessionError
Errors returned by ResumeSession

SendAutomationSignalError
Errors returned by SendAutomationSignal

SendCommandError
Errors returned by SendCommand

StartAssociationsOnceError
Errors returned by StartAssociationsOnce

StartAutomationExecutionError
Errors returned by StartAutomationExecution

StartSessionError
Errors returned by StartSession

StopAutomationExecutionError
Errors returned by StopAutomationExecution

TerminateSessionError
Errors returned by TerminateSession

UpdateAssociationError
Errors returned by UpdateAssociation

UpdateAssociationStatusError
Errors returned by UpdateAssociationStatus

UpdateDocumentDefaultVersionError
Errors returned by UpdateDocumentDefaultVersion

UpdateDocumentError
Errors returned by UpdateDocument

UpdateMaintenanceWindowError
Errors returned by UpdateMaintenanceWindow

UpdateMaintenanceWindowTargetError
Errors returned by UpdateMaintenanceWindowTarget

UpdateMaintenanceWindowTaskError
Errors returned by UpdateMaintenanceWindowTask

UpdateManagedInstanceRoleError
Errors returned by UpdateManagedInstanceRole

UpdateOpsItemError
Errors returned by UpdateOpsItem

UpdatePatchBaselineError
Errors returned by UpdatePatchBaseline

UpdateServiceSettingError
Errors returned by UpdateServiceSetting

Traits

Ssm
Trait representing the capabilities of the Amazon SSM API. Amazon SSM clients implement this trait.