Enum rusoto_ssm::PutParameterError

pub enum PutParameterError {
    HierarchyLevelLimitExceeded(String),
    HierarchyTypeMismatch(String),
    InternalServerError(String),
    InvalidAllowedPattern(String),
    InvalidKeyId(String),
    ParameterAlreadyExists(String),
    ParameterLimitExceeded(String),
    ParameterPatternMismatch(String),
    TooManyUpdates(String),
    UnsupportedParameterType(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by PutParameter

Variants
HierarchyLevelLimitExceeded(String)

A hierarchy can have a maximum of five levels. For example:

/Finance/Prod/IAD/OS/WinServ2016/license15

For more information, see Develop a Parameter Hierarchy.

HierarchyTypeMismatch(String)

Parameter Store does not support changing a parameter type in a hierarchy. For example, you can't change a parameter from a String type to a SecureString type. You must create a new, unique parameter.

InternalServerError(String)

An error occurred on the server side.

InvalidAllowedPattern(String)

The request does not meet the regular expression requirement.

InvalidKeyId(String)

The query key ID is not valid.

ParameterAlreadyExists(String)

The parameter already exists. You can't create duplicate parameters.

ParameterLimitExceeded(String)

You have exceeded the number of parameters for this AWS account. Delete one or more parameters and try again.

ParameterPatternMismatch(String)

The parameter name is not valid.

TooManyUpdates(String)

There are concurrent updates for a resource that supports one update at a time.

UnsupportedParameterType(String)

The parameter type is not supported.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl PutParameterError

fn from_body(body: &str) -> PutParameterError