Enum rusoto_ssm::SendCommandError

pub enum SendCommandError {
    DuplicateInstanceId(String),
    InternalServerError(String),
    InvalidDocument(String),
    InvalidInstanceId(String),
    InvalidNotificationConfig(String),
    InvalidOutputFolder(String),
    InvalidParameters(String),
    InvalidRole(String),
    MaxDocumentSizeExceeded(String),
    UnsupportedPlatformType(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by SendCommand

Variants
DuplicateInstanceId(String)

You cannot specify an instance ID in more than one association.

InternalServerError(String)

An error occurred on the server side.

InvalidDocument(String)

The specified document does not exist.

InvalidInstanceId(String)

The following problems can cause this exception:

You do not have permission to access the instance.

The SSM Agent is not running. On managed instances and Linux instances, verify that the SSM Agent is running. On EC2 Windows instances, verify that the EC2Config service is running.

The SSM Agent or EC2Config service is not registered to the SSM endpoint. Try reinstalling the SSM Agent or EC2Config service.

The instance is not in valid state. Valid states are: Running, Pending, Stopped, Stopping. Invalid states are: Shutting-down and Terminated.

InvalidNotificationConfig(String)

One or more configuration items is not valid. Verify that a valid Amazon Resource Name (ARN) was provided for an Amazon SNS topic.

InvalidOutputFolder(String)

The S3 bucket does not exist.

InvalidParameters(String)

You must specify values for all required parameters in the SSM document. You can only supply values to parameters defined in the SSM document.

InvalidRole(String)

The role name can't contain invalid characters. Also verify that you specified an IAM role for notifications that includes the required trust policy. For information about configuring the IAM role for Run Command notifications, see Configuring Amazon SNS Notifications for Run Command in the Amazon EC2 Systems Manager User Guide.

MaxDocumentSizeExceeded(String)

The size limit of a document is 64 KB.

UnsupportedPlatformType(String)

The document does not support the platform type of the given instance ID(s). For example, you sent an document for a Windows instance to a Linux instance.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl SendCommandError

fn from_body(body: &str) -> SendCommandError