Struct rusoto_ssm::AutomationExecution

pub struct AutomationExecution {
    pub automation_execution_id: Option<String>,
    pub automation_execution_status: Option<String>,
    pub document_name: Option<String>,
    pub document_version: Option<String>,
    pub execution_end_time: Option<f64>,
    pub execution_start_time: Option<f64>,
    pub failure_message: Option<String>,
    pub outputs: Option<HashMap<String, Vec<String>>>,
    pub parameters: Option<HashMap<String, Vec<String>>>,
    pub step_executions: Option<Vec<StepExecution>>,
}

Detailed information about the current state of an individual Automation execution.

Fields
automation_execution_id: Option<String>

The execution ID.

automation_execution_status: Option<String>

The execution status of the Automation.

document_name: Option<String>

The name of the Automation document used during the execution.

document_version: Option<String>

The version of the document to use during execution.

execution_end_time: Option<f64>

The time the execution finished.

execution_start_time: Option<f64>

The time the execution started.

failure_message: Option<String>

A message describing why an execution has failed, if the status is set to Failed.

outputs: Option<HashMap<String, Vec<String>>>

The list of execution outputs as defined in the automation document.

parameters: Option<HashMap<String, Vec<String>>>

The key-value map of execution parameters, which were supplied when calling StartAutomationExecution.

step_executions: Option<Vec<StepExecution>>

A list of details about the current state of all steps that comprise an execution. An Automation document contains a list of steps that are executed in order.