Struct rusoto_ssm::Command

pub struct Command {
    pub command_id: Option<String>,
    pub comment: Option<String>,
    pub completed_count: Option<i64>,
    pub document_name: Option<String>,
    pub error_count: Option<i64>,
    pub expires_after: Option<f64>,
    pub instance_ids: Option<Vec<String>>,
    pub max_concurrency: Option<String>,
    pub max_errors: Option<String>,
    pub notification_config: Option<NotificationConfig>,
    pub output_s3_bucket_name: Option<String>,
    pub output_s3_key_prefix: Option<String>,
    pub output_s3_region: Option<String>,
    pub parameters: Option<HashMap<String, Vec<String>>>,
    pub requested_date_time: Option<f64>,
    pub service_role: Option<String>,
    pub status: Option<String>,
    pub status_details: Option<String>,
    pub target_count: Option<i64>,
    pub targets: Option<Vec<Target>>,
}

Describes a command request.

Fields
command_id: Option<String>

A unique identifier for this command.

comment: Option<String>

User-specified information about the command, such as a brief description of what the command should do.

completed_count: Option<i64>

The number of targets for which the command invocation reached a terminal state. Terminal states include the following: Success, Failed, Execution Timed Out, Delivery Timed Out, Canceled, Terminated, or Undeliverable.

document_name: Option<String>

The name of the document requested for execution.

error_count: Option<i64>

The number of targets for which the status is Failed or Execution Timed Out.

expires_after: Option<f64>

If this time is reached and the command has not already started executing, it will not execute. Calculated based on the ExpiresAfter user input provided as part of the SendCommand API.

instance_ids: Option<Vec<String>>

The instance IDs against which this command was requested.

max_concurrency: Option<String>

The maximum number of instances that are allowed to execute the command at the same time. You can specify a number of instances, such as 10, or a percentage of instances, such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Executing a Command Using Systems Manager Run Command.

max_errors: Option<String>

The maximum number of errors allowed before the system stops sending the command to additional targets. You can specify a number of errors, such as 10, or a percentage or errors, such as 10%. The default value is 50. For more information about how to use MaxErrors, see Executing a Command Using Systems Manager Run Command.

notification_config: Option<NotificationConfig>

Configurations for sending notifications about command status changes.

output_s3_bucket_name: Option<String>

The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command.

output_s3_key_prefix: Option<String>

The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command.

output_s3_region: Option<String>

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon S3 bucket region.

parameters: Option<HashMap<String, Vec<String>>>

The parameter values to be inserted in the document when executing the command.

requested_date_time: Option<f64>

The date and time the command was requested.

service_role: Option<String>

The IAM service role that Run Command uses to act on your behalf when sending notifications about command status changes.

status: Option<String>

The status of the command.

status_details: Option<String>

A detailed status of the command execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Run Command Status. StatusDetails can be one of the following values:

Pending: The command has not been sent to any instances.

In Progress: The command has been sent to at least one instance but has not reached a final state on all instances.

Success: The command successfully executed on all invocations. This is a terminal state.

Delivery Timed Out: The value of MaxErrors or more command invocations shows a status of Delivery Timed Out. This is a terminal state.

Execution Timed Out: The value of MaxErrors or more command invocations shows a status of Execution Timed Out. This is a terminal state.

Failed: The value of MaxErrors or more command invocations shows a status of Failed. This is a terminal state.

Incomplete: The command was attempted on all instances and one or more invocations does not have a value of Success but not enough invocations failed for the status to be Failed. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Rate Exceeded: The number of instances targeted by the command exceeded the account limit for pending invocations. The system has canceled the command before executing it on any instance. This is a terminal state.

target_count: Option<i64>

The number of targets for the command.

targets: Option<Vec<Target>>

An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call.