Struct rusoto_ssm::GetMaintenanceWindowExecutionTaskResult

pub struct GetMaintenanceWindowExecutionTaskResult {
    pub end_time: Option<f64>,
    pub max_concurrency: Option<String>,
    pub max_errors: Option<String>,
    pub priority: Option<i64>,
    pub service_role: Option<String>,
    pub start_time: Option<f64>,
    pub status: Option<String>,
    pub status_details: Option<String>,
    pub task_arn: Option<String>,
    pub task_execution_id: Option<String>,
    pub task_parameters: Option<Vec<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>>,
    pub type_: Option<String>,
    pub window_execution_id: Option<String>,
}

Fields
end_time: Option<f64>

The time the task execution completed.

max_concurrency: Option<String>

The defined maximum number of task executions that could be run in parallel.

max_errors: Option<String>

The defined maximum number of task execution errors allowed before scheduling of the task execution would have been stopped.

priority: Option<i64>

The priority of the task.

service_role: Option<String>

The role that was assumed when executing the task.

start_time: Option<f64>

The time the task execution started.

status: Option<String>

The status of the task.

status_details: Option<String>

The details explaining the Status. Only available for certain status values.

task_arn: Option<String>

The ARN of the executed task.

task_execution_id: Option<String>

The ID of the specific task execution in the Maintenance Window task that was retrieved.

task_parameters: Option<Vec<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>>

The parameters passed to the task when it was executed. The map has the following format:

Key: string, between 1 and 255 characters

Value: an array of strings, each string is between 1 and 255 characters

type_: Option<String>

The type of task executed.

window_execution_id: Option<String>

The ID of the Maintenance Window execution that includes the task.