Struct rusoto_ssm::InstancePatchState

pub struct InstancePatchState {
    pub baseline_id: String,
    pub failed_count: Option<i64>,
    pub installed_count: Option<i64>,
    pub installed_other_count: Option<i64>,
    pub instance_id: String,
    pub missing_count: Option<i64>,
    pub not_applicable_count: Option<i64>,
    pub operation: String,
    pub operation_end_time: f64,
    pub operation_start_time: f64,
    pub owner_information: Option<String>,
    pub patch_group: String,
    pub snapshot_id: Option<String>,
}

Defines the high-level patch compliance state for a managed instance, providing information about the number of installed, missing, not applicable, and failed patches along with metadata about the operation when this information was gathered for the instance.

Fields
baseline_id: String

The ID of the patch baseline used to patch the instance.

failed_count: Option<i64>

The number of patches from the patch baseline that were attempted to be installed during the last patching operation, but failed to install.

installed_count: Option<i64>

The number of patches from the patch baseline that are installed on the instance.

installed_other_count: Option<i64>

The number of patches not specified in the patch baseline that are installed on the instance.

instance_id: String

The ID of the managed instance the high-level patch compliance information was collected for.

missing_count: Option<i64>

The number of patches from the patch baseline that are applicable for the instance but aren't currently installed.

not_applicable_count: Option<i64>

The number of patches from the patch baseline that aren't applicable for the instance and hence aren't installed on the instance.

operation: String

The type of patching operation that was performed: SCAN (assess patch compliance state) or INSTALL (install missing patches).

operation_end_time: f64

The time the most recent patching operation completed on the instance.

operation_start_time: f64

The time the most recent patching operation was started on the instance.

owner_information: Option<String>

Placeholder information, this field will always be empty in the current release of the service.

patch_group: String

The name of the patch group the managed instance belongs to.

snapshot_id: Option<String>

The ID of the patch baseline snapshot used during the patching operation when this compliance data was collected.