Struct rusoto_ssm::MaintenanceWindowExecutionTaskInvocationIdentity

pub struct MaintenanceWindowExecutionTaskInvocationIdentity {
    pub end_time: Option<f64>,
    pub execution_id: Option<String>,
    pub invocation_id: Option<String>,
    pub owner_information: Option<String>,
    pub parameters: Option<String>,
    pub start_time: Option<f64>,
    pub status: Option<String>,
    pub status_details: Option<String>,
    pub task_execution_id: Option<String>,
    pub window_execution_id: Option<String>,
    pub window_target_id: Option<String>,
}

Describes the information about a task invocation for a particular target as part of a task execution performed as part of a Maintenance Window execution.

Fields
end_time: Option<f64>

The time the invocation finished.

execution_id: Option<String>

The ID of the action performed in the service that actually handled the task invocation. If the task type is RUN_COMMAND, this value is the command ID.

invocation_id: Option<String>

The ID of the task invocation.

owner_information: Option<String>

User-provided value that was specified when the target was registered with the Maintenance Window. This was also included in any CloudWatch events raised during the task invocation.

parameters: Option<String>

The parameters that were provided for the invocation when it was executed.

start_time: Option<f64>

The time the invocation started.

status: Option<String>

The status of the task invocation.

status_details: Option<String>

The details explaining the status of the task invocation. Only available for certain Status values.

task_execution_id: Option<String>

The ID of the specific task execution in the Maintenance Window execution.

window_execution_id: Option<String>

The ID of the Maintenance Window execution that ran the task.

window_target_id: Option<String>

The ID of the target definition in this Maintenance Window the invocation was performed for.