Struct rusoto_ssm::RegisterTaskWithMaintenanceWindowRequest

pub struct RegisterTaskWithMaintenanceWindowRequest {
    pub client_token: Option<String>,
    pub logging_info: Option<LoggingInfo>,
    pub max_concurrency: String,
    pub max_errors: String,
    pub priority: Option<i64>,
    pub service_role_arn: String,
    pub targets: Vec<Target>,
    pub task_arn: String,
    pub task_parameters: Option<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>,
    pub task_type: String,
    pub window_id: String,
}

Fields
client_token: Option<String>

User-provided idempotency token.

logging_info: Option<LoggingInfo>

A structure containing information about an Amazon S3 bucket to write instance-level logs to.

max_concurrency: String

The maximum number of targets this task can be run for in parallel.

max_errors: String

The maximum number of errors allowed before this task stops being scheduled.

priority: Option<i64>

The priority of the task in the Maintenance Window, the lower the number the higher the priority. Tasks in a Maintenance Window are scheduled in priority order with tasks that have the same priority scheduled in parallel.

service_role_arn: String

The role that should be assumed when executing the task.

targets: Vec<Target>

The targets (either instances or tags). Instances are specified using Key=instanceids,Values=<instanceid1>,<instanceid2>. Tags are specified using Key=<tag name>,Values=<tag value>.

task_arn: String

The ARN of the task to execute

task_parameters: Option<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>

The parameters that should be passed to the task when it is executed.

task_type: String

The type of task being registered.

window_id: String

The id of the Maintenance Window the task should be added to.