Struct rusoto_ssm::UpdateMaintenanceWindowTaskRequest

pub struct UpdateMaintenanceWindowTaskRequest {
    pub description: Option<String>,
    pub logging_info: Option<LoggingInfo>,
    pub max_concurrency: Option<String>,
    pub max_errors: Option<String>,
    pub name: Option<String>,
    pub priority: Option<i64>,
    pub replace: Option<bool>,
    pub service_role_arn: Option<String>,
    pub targets: Option<Vec<Target>>,
    pub task_arn: Option<String>,
    pub task_invocation_parameters: Option<MaintenanceWindowTaskInvocationParameters>,
    pub task_parameters: Option<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>,
    pub window_id: String,
    pub window_task_id: String,
}

Fields
description: Option<String>

The new task description to specify.

logging_info: Option<LoggingInfo>

The new logging location in Amazon S3 to specify.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

max_concurrency: Option<String>

The new MaxConcurrency value you want to specify. MaxConcurrency is the number of targets that are allowed to run this task in parallel.

max_errors: Option<String>

The new MaxErrors value to specify. MaxErrors is the maximum number of errors that are allowed before the task stops being scheduled.

name: Option<String>

The new task name to specify.

priority: Option<i64>

The new task priority to specify. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

replace: Option<bool>

If True, then all fields that are required by the RegisterTaskWithMaintenanceWndow action are also required for this API request. Optional fields that are not specified are set to null.

service_role_arn: Option<String>

The ARN of the IAM service role for Systems Manager to assume when running a maintenance window task. If you do not specify a service role ARN, Systems Manager uses your account's service-linked role. If no service-linked role for Systems Manager exists in your account, it is created when you run RegisterTaskWithMaintenanceWindow.

For more information, see the following topics in the in the AWS Systems Manager User Guide:

Service-Linked Role Permissions for Systems Manager

Should I Use a Service-Linked Role or a Custom Service Role to Run Maintenance Window Tasks?

targets: Option<Vec<Target>>

The targets (either instances or tags) to modify. Instances are specified using Key=instanceids,Values=instanceID_1,instanceID_2. Tags are specified using Key=tag_name,Values=tag_value.

task_arn: Option<String>

The task ARN to modify.

task_invocation_parameters: Option<MaintenanceWindowTaskInvocationParameters>

The parameters that the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

task_parameters: Option<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>

The parameters to modify.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

The map has the following format:

Key: string, between 1 and 255 characters

Value: an array of strings, each string is between 1 and 255 characters

window_id: String

The maintenance window ID that contains the task to modify.

window_task_id: String

The task ID to modify.