Trait rusoto_ssm::Ssm

pub trait Ssm {
    fn add_tags_to_resource(
        &self, 
        input: &AddTagsToResourceRequest
    ) -> Result<AddTagsToResourceResult, AddTagsToResourceError>;
    fn cancel_command(
        &self, 
        input: &CancelCommandRequest
    ) -> Result<CancelCommandResult, CancelCommandError>;
    fn create_activation(
        &self, 
        input: &CreateActivationRequest
    ) -> Result<CreateActivationResult, CreateActivationError>;
    fn create_association(
        &self, 
        input: &CreateAssociationRequest
    ) -> Result<CreateAssociationResult, CreateAssociationError>;
    fn create_association_batch(
        &self, 
        input: &CreateAssociationBatchRequest
    ) -> Result<CreateAssociationBatchResult, CreateAssociationBatchError>;
    fn create_document(
        &self, 
        input: &CreateDocumentRequest
    ) -> Result<CreateDocumentResult, CreateDocumentError>;
    fn create_maintenance_window(
        &self, 
        input: &CreateMaintenanceWindowRequest
    ) -> Result<CreateMaintenanceWindowResult, CreateMaintenanceWindowError>;
    fn create_patch_baseline(
        &self, 
        input: &CreatePatchBaselineRequest
    ) -> Result<CreatePatchBaselineResult, CreatePatchBaselineError>;
    fn delete_activation(
        &self, 
        input: &DeleteActivationRequest
    ) -> Result<DeleteActivationResult, DeleteActivationError>;
    fn delete_association(
        &self, 
        input: &DeleteAssociationRequest
    ) -> Result<DeleteAssociationResult, DeleteAssociationError>;
    fn delete_document(
        &self, 
        input: &DeleteDocumentRequest
    ) -> Result<DeleteDocumentResult, DeleteDocumentError>;
    fn delete_maintenance_window(
        &self, 
        input: &DeleteMaintenanceWindowRequest
    ) -> Result<DeleteMaintenanceWindowResult, DeleteMaintenanceWindowError>;
    fn delete_parameter(
        &self, 
        input: &DeleteParameterRequest
    ) -> Result<DeleteParameterResult, DeleteParameterError>;
    fn delete_parameters(
        &self, 
        input: &DeleteParametersRequest
    ) -> Result<DeleteParametersResult, DeleteParametersError>;
    fn delete_patch_baseline(
        &self, 
        input: &DeletePatchBaselineRequest
    ) -> Result<DeletePatchBaselineResult, DeletePatchBaselineError>;
    fn deregister_managed_instance(
        &self, 
        input: &DeregisterManagedInstanceRequest
    ) -> Result<DeregisterManagedInstanceResult, DeregisterManagedInstanceError>;
    fn deregister_patch_baseline_for_patch_group(
        &self, 
        input: &DeregisterPatchBaselineForPatchGroupRequest
    ) -> Result<DeregisterPatchBaselineForPatchGroupResult, DeregisterPatchBaselineForPatchGroupError>;
    fn deregister_target_from_maintenance_window(
        &self, 
        input: &DeregisterTargetFromMaintenanceWindowRequest
    ) -> Result<DeregisterTargetFromMaintenanceWindowResult, DeregisterTargetFromMaintenanceWindowError>;
    fn deregister_task_from_maintenance_window(
        &self, 
        input: &DeregisterTaskFromMaintenanceWindowRequest
    ) -> Result<DeregisterTaskFromMaintenanceWindowResult, DeregisterTaskFromMaintenanceWindowError>;
    fn describe_activations(
        &self, 
        input: &DescribeActivationsRequest
    ) -> Result<DescribeActivationsResult, DescribeActivationsError>;
    fn describe_association(
        &self, 
        input: &DescribeAssociationRequest
    ) -> Result<DescribeAssociationResult, DescribeAssociationError>;
    fn describe_automation_executions(
        &self, 
        input: &DescribeAutomationExecutionsRequest
    ) -> Result<DescribeAutomationExecutionsResult, DescribeAutomationExecutionsError>;
    fn describe_available_patches(
        &self, 
        input: &DescribeAvailablePatchesRequest
    ) -> Result<DescribeAvailablePatchesResult, DescribeAvailablePatchesError>;
    fn describe_document(
        &self, 
        input: &DescribeDocumentRequest
    ) -> Result<DescribeDocumentResult, DescribeDocumentError>;
    fn describe_document_permission(
        &self, 
        input: &DescribeDocumentPermissionRequest
    ) -> Result<DescribeDocumentPermissionResponse, DescribeDocumentPermissionError>;
    fn describe_effective_instance_associations(
        &self, 
        input: &DescribeEffectiveInstanceAssociationsRequest
    ) -> Result<DescribeEffectiveInstanceAssociationsResult, DescribeEffectiveInstanceAssociationsError>;
    fn describe_effective_patches_for_patch_baseline(
        &self, 
        input: &DescribeEffectivePatchesForPatchBaselineRequest
    ) -> Result<DescribeEffectivePatchesForPatchBaselineResult, DescribeEffectivePatchesForPatchBaselineError>;
    fn describe_instance_associations_status(
        &self, 
        input: &DescribeInstanceAssociationsStatusRequest
    ) -> Result<DescribeInstanceAssociationsStatusResult, DescribeInstanceAssociationsStatusError>;
    fn describe_instance_information(
        &self, 
        input: &DescribeInstanceInformationRequest
    ) -> Result<DescribeInstanceInformationResult, DescribeInstanceInformationError>;
    fn describe_instance_patch_states(
        &self, 
        input: &DescribeInstancePatchStatesRequest
    ) -> Result<DescribeInstancePatchStatesResult, DescribeInstancePatchStatesError>;
    fn describe_instance_patch_states_for_patch_group(
        &self, 
        input: &DescribeInstancePatchStatesForPatchGroupRequest
    ) -> Result<DescribeInstancePatchStatesForPatchGroupResult, DescribeInstancePatchStatesForPatchGroupError>;
    fn describe_instance_patches(
        &self, 
        input: &DescribeInstancePatchesRequest
    ) -> Result<DescribeInstancePatchesResult, DescribeInstancePatchesError>;
    fn describe_maintenance_window_execution_task_invocations(
        &self, 
        input: &DescribeMaintenanceWindowExecutionTaskInvocationsRequest
    ) -> Result<DescribeMaintenanceWindowExecutionTaskInvocationsResult, DescribeMaintenanceWindowExecutionTaskInvocationsError>;
    fn describe_maintenance_window_execution_tasks(
        &self, 
        input: &DescribeMaintenanceWindowExecutionTasksRequest
    ) -> Result<DescribeMaintenanceWindowExecutionTasksResult, DescribeMaintenanceWindowExecutionTasksError>;
    fn describe_maintenance_window_executions(
        &self, 
        input: &DescribeMaintenanceWindowExecutionsRequest
    ) -> Result<DescribeMaintenanceWindowExecutionsResult, DescribeMaintenanceWindowExecutionsError>;
    fn describe_maintenance_window_targets(
        &self, 
        input: &DescribeMaintenanceWindowTargetsRequest
    ) -> Result<DescribeMaintenanceWindowTargetsResult, DescribeMaintenanceWindowTargetsError>;
    fn describe_maintenance_window_tasks(
        &self, 
        input: &DescribeMaintenanceWindowTasksRequest
    ) -> Result<DescribeMaintenanceWindowTasksResult, DescribeMaintenanceWindowTasksError>;
    fn describe_maintenance_windows(
        &self, 
        input: &DescribeMaintenanceWindowsRequest
    ) -> Result<DescribeMaintenanceWindowsResult, DescribeMaintenanceWindowsError>;
    fn describe_parameters(
        &self, 
        input: &DescribeParametersRequest
    ) -> Result<DescribeParametersResult, DescribeParametersError>;
    fn describe_patch_baselines(
        &self, 
        input: &DescribePatchBaselinesRequest
    ) -> Result<DescribePatchBaselinesResult, DescribePatchBaselinesError>;
    fn describe_patch_group_state(
        &self, 
        input: &DescribePatchGroupStateRequest
    ) -> Result<DescribePatchGroupStateResult, DescribePatchGroupStateError>;
    fn describe_patch_groups(
        &self, 
        input: &DescribePatchGroupsRequest
    ) -> Result<DescribePatchGroupsResult, DescribePatchGroupsError>;
    fn get_automation_execution(
        &self, 
        input: &GetAutomationExecutionRequest
    ) -> Result<GetAutomationExecutionResult, GetAutomationExecutionError>;
    fn get_command_invocation(
        &self, 
        input: &GetCommandInvocationRequest
    ) -> Result<GetCommandInvocationResult, GetCommandInvocationError>;
    fn get_default_patch_baseline(
        &self
    ) -> Result<GetDefaultPatchBaselineResult, GetDefaultPatchBaselineError>;
    fn get_deployable_patch_snapshot_for_instance(
        &self, 
        input: &GetDeployablePatchSnapshotForInstanceRequest
    ) -> Result<GetDeployablePatchSnapshotForInstanceResult, GetDeployablePatchSnapshotForInstanceError>;
    fn get_document(
        &self, 
        input: &GetDocumentRequest
    ) -> Result<GetDocumentResult, GetDocumentError>;
    fn get_inventory(
        &self, 
        input: &GetInventoryRequest
    ) -> Result<GetInventoryResult, GetInventoryError>;
    fn get_inventory_schema(
        &self, 
        input: &GetInventorySchemaRequest
    ) -> Result<GetInventorySchemaResult, GetInventorySchemaError>;
    fn get_maintenance_window(
        &self, 
        input: &GetMaintenanceWindowRequest
    ) -> Result<GetMaintenanceWindowResult, GetMaintenanceWindowError>;
    fn get_maintenance_window_execution(
        &self, 
        input: &GetMaintenanceWindowExecutionRequest
    ) -> Result<GetMaintenanceWindowExecutionResult, GetMaintenanceWindowExecutionError>;
    fn get_maintenance_window_execution_task(
        &self, 
        input: &GetMaintenanceWindowExecutionTaskRequest
    ) -> Result<GetMaintenanceWindowExecutionTaskResult, GetMaintenanceWindowExecutionTaskError>;
    fn get_parameter(
        &self, 
        input: &GetParameterRequest
    ) -> Result<GetParameterResult, GetParameterError>;
    fn get_parameter_history(
        &self, 
        input: &GetParameterHistoryRequest
    ) -> Result<GetParameterHistoryResult, GetParameterHistoryError>;
    fn get_parameters(
        &self, 
        input: &GetParametersRequest
    ) -> Result<GetParametersResult, GetParametersError>;
    fn get_parameters_by_path(
        &self, 
        input: &GetParametersByPathRequest
    ) -> Result<GetParametersByPathResult, GetParametersByPathError>;
    fn get_patch_baseline(
        &self, 
        input: &GetPatchBaselineRequest
    ) -> Result<GetPatchBaselineResult, GetPatchBaselineError>;
    fn get_patch_baseline_for_patch_group(
        &self, 
        input: &GetPatchBaselineForPatchGroupRequest
    ) -> Result<GetPatchBaselineForPatchGroupResult, GetPatchBaselineForPatchGroupError>;
    fn list_associations(
        &self, 
        input: &ListAssociationsRequest
    ) -> Result<ListAssociationsResult, ListAssociationsError>;
    fn list_command_invocations(
        &self, 
        input: &ListCommandInvocationsRequest
    ) -> Result<ListCommandInvocationsResult, ListCommandInvocationsError>;
    fn list_commands(
        &self, 
        input: &ListCommandsRequest
    ) -> Result<ListCommandsResult, ListCommandsError>;
    fn list_document_versions(
        &self, 
        input: &ListDocumentVersionsRequest
    ) -> Result<ListDocumentVersionsResult, ListDocumentVersionsError>;
    fn list_documents(
        &self, 
        input: &ListDocumentsRequest
    ) -> Result<ListDocumentsResult, ListDocumentsError>;
    fn list_inventory_entries(
        &self, 
        input: &ListInventoryEntriesRequest
    ) -> Result<ListInventoryEntriesResult, ListInventoryEntriesError>;
    fn list_tags_for_resource(
        &self, 
        input: &ListTagsForResourceRequest
    ) -> Result<ListTagsForResourceResult, ListTagsForResourceError>;
    fn modify_document_permission(
        &self, 
        input: &ModifyDocumentPermissionRequest
    ) -> Result<ModifyDocumentPermissionResponse, ModifyDocumentPermissionError>;
    fn put_inventory(
        &self, 
        input: &PutInventoryRequest
    ) -> Result<PutInventoryResult, PutInventoryError>;
    fn put_parameter(
        &self, 
        input: &PutParameterRequest
    ) -> Result<PutParameterResult, PutParameterError>;
    fn register_default_patch_baseline(
        &self, 
        input: &RegisterDefaultPatchBaselineRequest
    ) -> Result<RegisterDefaultPatchBaselineResult, RegisterDefaultPatchBaselineError>;
    fn register_patch_baseline_for_patch_group(
        &self, 
        input: &RegisterPatchBaselineForPatchGroupRequest
    ) -> Result<RegisterPatchBaselineForPatchGroupResult, RegisterPatchBaselineForPatchGroupError>;
    fn register_target_with_maintenance_window(
        &self, 
        input: &RegisterTargetWithMaintenanceWindowRequest
    ) -> Result<RegisterTargetWithMaintenanceWindowResult, RegisterTargetWithMaintenanceWindowError>;
    fn register_task_with_maintenance_window(
        &self, 
        input: &RegisterTaskWithMaintenanceWindowRequest
    ) -> Result<RegisterTaskWithMaintenanceWindowResult, RegisterTaskWithMaintenanceWindowError>;
    fn remove_tags_from_resource(
        &self, 
        input: &RemoveTagsFromResourceRequest
    ) -> Result<RemoveTagsFromResourceResult, RemoveTagsFromResourceError>;
    fn send_command(
        &self, 
        input: &SendCommandRequest
    ) -> Result<SendCommandResult, SendCommandError>;
    fn start_automation_execution(
        &self, 
        input: &StartAutomationExecutionRequest
    ) -> Result<StartAutomationExecutionResult, StartAutomationExecutionError>;
    fn stop_automation_execution(
        &self, 
        input: &StopAutomationExecutionRequest
    ) -> Result<StopAutomationExecutionResult, StopAutomationExecutionError>;
    fn update_association(
        &self, 
        input: &UpdateAssociationRequest
    ) -> Result<UpdateAssociationResult, UpdateAssociationError>;
    fn update_association_status(
        &self, 
        input: &UpdateAssociationStatusRequest
    ) -> Result<UpdateAssociationStatusResult, UpdateAssociationStatusError>;
    fn update_document(
        &self, 
        input: &UpdateDocumentRequest
    ) -> Result<UpdateDocumentResult, UpdateDocumentError>;
    fn update_document_default_version(
        &self, 
        input: &UpdateDocumentDefaultVersionRequest
    ) -> Result<UpdateDocumentDefaultVersionResult, UpdateDocumentDefaultVersionError>;
    fn update_maintenance_window(
        &self, 
        input: &UpdateMaintenanceWindowRequest
    ) -> Result<UpdateMaintenanceWindowResult, UpdateMaintenanceWindowError>;
    fn update_managed_instance_role(
        &self, 
        input: &UpdateManagedInstanceRoleRequest
    ) -> Result<UpdateManagedInstanceRoleResult, UpdateManagedInstanceRoleError>;
    fn update_patch_baseline(
        &self, 
        input: &UpdatePatchBaselineRequest
    ) -> Result<UpdatePatchBaselineResult, UpdatePatchBaselineError>;
}

Trait representing the capabilities of the Amazon SSM API. Amazon SSM clients implement this trait.

Required Methods

fn add_tags_to_resource(
&self,
input: &AddTagsToResourceRequest
) -> Result<AddTagsToResourceResult, AddTagsToResourceError>

Adds or overwrites one or more tags for the specified resource. Tags are metadata that you assign to your managed instances, Maintenance Windows, or Parameter Store parameters. Tags enable you to categorize your resources in different ways, for example, by purpose, owner, or environment. Each tag consists of a key and an optional value, both of which you define. For example, you could define a set of tags for your account's managed instances that helps you track each instance's owner and stack level. For example: Key=Owner and Value=DbAdmin, SysAdmin, or Dev. Or Key=Stack and Value=Production, Pre-Production, or Test.

Each resource can have a maximum of 10 tags.

We recommend that you devise a set of tag keys that meets your needs for each resource type. Using a consistent set of tag keys makes it easier for you to manage your resources. You can search and filter the resources based on the tags you add. Tags don't have any semantic meaning to Amazon EC2 and are interpreted strictly as a string of characters.

For more information about tags, see Tagging Your Amazon EC2 Resources in the Amazon EC2 User Guide.

fn cancel_command(
&self,
input: &CancelCommandRequest
) -> Result<CancelCommandResult, CancelCommandError>

Attempts to cancel the command specified by the Command ID. There is no guarantee that the command will be terminated and the underlying process stopped.

fn create_activation(
&self,
input: &CreateActivationRequest
) -> Result<CreateActivationResult, CreateActivationError>

Registers your on-premises server or virtual machine with Amazon EC2 so that you can manage these resources using Run Command. An on-premises server or virtual machine that has been registered with EC2 is called a managed instance. For more information about activations, see Setting Up Systems Manager in Hybrid Environments.

fn create_association(
&self,
input: &CreateAssociationRequest
) -> Result<CreateAssociationResult, CreateAssociationError>

Associates the specified Systems Manager document with the specified instances or targets.

When you associate a document with one or more instances using instance IDs or tags, the SSM Agent running on the instance processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system throws the AssociationAlreadyExists exception.

fn create_association_batch(
&self,
input: &CreateAssociationBatchRequest
) -> Result<CreateAssociationBatchResult, CreateAssociationBatchError>

Associates the specified Systems Manager document with the specified instances or targets.

When you associate a document with one or more instances using instance IDs or tags, the SSM Agent running on the instance processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system throws the AssociationAlreadyExists exception.

fn create_document(
&self,
input: &CreateDocumentRequest
) -> Result<CreateDocumentResult, CreateDocumentError>

Creates a Systems Manager document.

After you create a document, you can use CreateAssociation to associate it with one or more running instances.

fn create_maintenance_window(
&self,
input: &CreateMaintenanceWindowRequest
) -> Result<CreateMaintenanceWindowResult, CreateMaintenanceWindowError>

Creates a new Maintenance Window.

fn create_patch_baseline(
&self,
input: &CreatePatchBaselineRequest
) -> Result<CreatePatchBaselineResult, CreatePatchBaselineError>

Creates a patch baseline.

fn delete_activation(
&self,
input: &DeleteActivationRequest
) -> Result<DeleteActivationResult, DeleteActivationError>

Deletes an activation. You are not required to delete an activation. If you delete an activation, you can no longer use it to register additional managed instances. Deleting an activation does not de-register managed instances. You must manually de-register managed instances.

fn delete_association(
&self,
input: &DeleteAssociationRequest
) -> Result<DeleteAssociationResult, DeleteAssociationError>

Disassociates the specified Systems Manager document from the specified instance.

When you disassociate a document from an instance, it does not change the configuration of the instance. To change the configuration state of an instance after you disassociate a document, you must create a new document with the desired configuration and associate it with the instance.

fn delete_document(
&self,
input: &DeleteDocumentRequest
) -> Result<DeleteDocumentResult, DeleteDocumentError>

Deletes the Systems Manager document and all instance associations to the document.

Before you delete the document, we recommend that you use DeleteAssociation to disassociate all instances that are associated with the document.

fn delete_maintenance_window(
&self,
input: &DeleteMaintenanceWindowRequest
) -> Result<DeleteMaintenanceWindowResult, DeleteMaintenanceWindowError>

Deletes a Maintenance Window.

fn delete_parameter(
&self,
input: &DeleteParameterRequest
) -> Result<DeleteParameterResult, DeleteParameterError>

Delete a parameter from the system.

fn delete_parameters(
&self,
input: &DeleteParametersRequest
) -> Result<DeleteParametersResult, DeleteParametersError>

Delete a list of parameters.

fn delete_patch_baseline(
&self,
input: &DeletePatchBaselineRequest
) -> Result<DeletePatchBaselineResult, DeletePatchBaselineError>

Deletes a patch baseline.

fn deregister_managed_instance(
&self,
input: &DeregisterManagedInstanceRequest
) -> Result<DeregisterManagedInstanceResult, DeregisterManagedInstanceError>

Removes the server or virtual machine from the list of registered servers. You can reregister the instance again at any time. If you don't plan to use Run Command on the server, we suggest uninstalling the SSM Agent first.

fn deregister_patch_baseline_for_patch_group(
&self,
input: &DeregisterPatchBaselineForPatchGroupRequest
) -> Result<DeregisterPatchBaselineForPatchGroupResult, DeregisterPatchBaselineForPatchGroupError>

Removes a patch group from a patch baseline.

fn deregister_target_from_maintenance_window(
&self,
input: &DeregisterTargetFromMaintenanceWindowRequest
) -> Result<DeregisterTargetFromMaintenanceWindowResult, DeregisterTargetFromMaintenanceWindowError>

Removes a target from a Maintenance Window.

fn deregister_task_from_maintenance_window(
&self,
input: &DeregisterTaskFromMaintenanceWindowRequest
) -> Result<DeregisterTaskFromMaintenanceWindowResult, DeregisterTaskFromMaintenanceWindowError>

Removes a task from a Maintenance Window.

fn describe_activations(
&self,
input: &DescribeActivationsRequest
) -> Result<DescribeActivationsResult, DescribeActivationsError>

Details about the activation, including: the date and time the activation was created, the expiration date, the IAM role assigned to the instances in the activation, and the number of instances activated by this registration.

fn describe_association(
&self,
input: &DescribeAssociationRequest
) -> Result<DescribeAssociationResult, DescribeAssociationError>

Describes the associations for the specified Systems Manager document or instance.

fn describe_automation_executions(
&self,
input: &DescribeAutomationExecutionsRequest
) -> Result<DescribeAutomationExecutionsResult, DescribeAutomationExecutionsError>

Provides details about all active and terminated Automation executions.

fn describe_available_patches(
&self,
input: &DescribeAvailablePatchesRequest
) -> Result<DescribeAvailablePatchesResult, DescribeAvailablePatchesError>

Lists all patches that could possibly be included in a patch baseline.

fn describe_document(
&self,
input: &DescribeDocumentRequest
) -> Result<DescribeDocumentResult, DescribeDocumentError>

Describes the specified SSM document.

fn describe_document_permission(
&self,
input: &DescribeDocumentPermissionRequest
) -> Result<DescribeDocumentPermissionResponse, DescribeDocumentPermissionError>

Describes the permissions for a Systems Manager document. If you created the document, you are the owner. If a document is shared, it can either be shared privately (by specifying a user's AWS account ID) or publicly (All).

fn describe_effective_instance_associations(
&self,
input: &DescribeEffectiveInstanceAssociationsRequest
) -> Result<DescribeEffectiveInstanceAssociationsResult, DescribeEffectiveInstanceAssociationsError>

All associations for the instance(s).

fn describe_effective_patches_for_patch_baseline(
&self,
input: &DescribeEffectivePatchesForPatchBaselineRequest
) -> Result<DescribeEffectivePatchesForPatchBaselineResult, DescribeEffectivePatchesForPatchBaselineError>

Retrieves the current effective patches (the patch and the approval state) for the specified patch baseline.

fn describe_instance_associations_status(
&self,
input: &DescribeInstanceAssociationsStatusRequest
) -> Result<DescribeInstanceAssociationsStatusResult, DescribeInstanceAssociationsStatusError>

The status of the associations for the instance(s).

fn describe_instance_information(
&self,
input: &DescribeInstanceInformationRequest
) -> Result<DescribeInstanceInformationResult, DescribeInstanceInformationError>

Describes one or more of your instances. You can use this to get information about instances like the operating system platform, the SSM Agent version (Linux), status etc. If you specify one or more instance IDs, it returns information for those instances. If you do not specify instance IDs, it returns information for all your instances. If you specify an instance ID that is not valid or an instance that you do not own, you receive an error.

fn describe_instance_patch_states(
&self,
input: &DescribeInstancePatchStatesRequest
) -> Result<DescribeInstancePatchStatesResult, DescribeInstancePatchStatesError>

Retrieves the high-level patch state of one or more instances.

fn describe_instance_patch_states_for_patch_group(
&self,
input: &DescribeInstancePatchStatesForPatchGroupRequest
) -> Result<DescribeInstancePatchStatesForPatchGroupResult, DescribeInstancePatchStatesForPatchGroupError>

Retrieves the high-level patch state for the instances in the specified patch group.

fn describe_instance_patches(
&self,
input: &DescribeInstancePatchesRequest
) -> Result<DescribeInstancePatchesResult, DescribeInstancePatchesError>

Retrieves information about the patches on the specified instance and their state relative to the patch baseline being used for the instance.

fn describe_maintenance_window_execution_task_invocations(
&self,
input: &DescribeMaintenanceWindowExecutionTaskInvocationsRequest
) -> Result<DescribeMaintenanceWindowExecutionTaskInvocationsResult, DescribeMaintenanceWindowExecutionTaskInvocationsError>

Retrieves the individual task executions (one per target) for a particular task executed as part of a Maintenance Window execution.

fn describe_maintenance_window_execution_tasks(
&self,
input: &DescribeMaintenanceWindowExecutionTasksRequest
) -> Result<DescribeMaintenanceWindowExecutionTasksResult, DescribeMaintenanceWindowExecutionTasksError>

For a given Maintenance Window execution, lists the tasks that were executed.

fn describe_maintenance_window_executions(
&self,
input: &DescribeMaintenanceWindowExecutionsRequest
) -> Result<DescribeMaintenanceWindowExecutionsResult, DescribeMaintenanceWindowExecutionsError>

Lists the executions of a Maintenance Window (meaning, information about when the Maintenance Window was scheduled to be active and information about tasks registered and run with the Maintenance Window).

fn describe_maintenance_window_targets(
&self,
input: &DescribeMaintenanceWindowTargetsRequest
) -> Result<DescribeMaintenanceWindowTargetsResult, DescribeMaintenanceWindowTargetsError>

Lists the targets registered with the Maintenance Window.

fn describe_maintenance_window_tasks(
&self,
input: &DescribeMaintenanceWindowTasksRequest
) -> Result<DescribeMaintenanceWindowTasksResult, DescribeMaintenanceWindowTasksError>

Lists the tasks in a Maintenance Window.

fn describe_maintenance_windows(
&self,
input: &DescribeMaintenanceWindowsRequest
) -> Result<DescribeMaintenanceWindowsResult, DescribeMaintenanceWindowsError>

Retrieves the Maintenance Windows in an AWS account.

fn describe_parameters(
&self,
input: &DescribeParametersRequest
) -> Result<DescribeParametersResult, DescribeParametersError>

Get information about a parameter.

fn describe_patch_baselines(
&self,
input: &DescribePatchBaselinesRequest
) -> Result<DescribePatchBaselinesResult, DescribePatchBaselinesError>

Lists the patch baselines in your AWS account.

fn describe_patch_group_state(
&self,
input: &DescribePatchGroupStateRequest
) -> Result<DescribePatchGroupStateResult, DescribePatchGroupStateError>

Returns high-level aggregated patch compliance state for a patch group.

fn describe_patch_groups(
&self,
input: &DescribePatchGroupsRequest
) -> Result<DescribePatchGroupsResult, DescribePatchGroupsError>

Lists all patch groups that have been registered with patch baselines.

fn get_automation_execution(
&self,
input: &GetAutomationExecutionRequest
) -> Result<GetAutomationExecutionResult, GetAutomationExecutionError>

Get detailed information about a particular Automation execution.

fn get_command_invocation(
&self,
input: &GetCommandInvocationRequest
) -> Result<GetCommandInvocationResult, GetCommandInvocationError>

Returns detailed information about command execution for an invocation or plugin.

fn get_default_patch_baseline(
&self
) -> Result<GetDefaultPatchBaselineResult, GetDefaultPatchBaselineError>

Retrieves the default patch baseline.

fn get_deployable_patch_snapshot_for_instance(
&self,
input: &GetDeployablePatchSnapshotForInstanceRequest
) -> Result<GetDeployablePatchSnapshotForInstanceResult, GetDeployablePatchSnapshotForInstanceError>

Retrieves the current snapshot for the patch baseline the instance uses. This API is primarily used by the AWS-ApplyPatchBaseline Systems Manager document.

fn get_document(
&self,
input: &GetDocumentRequest
) -> Result<GetDocumentResult, GetDocumentError>

Gets the contents of the specified SSM document.

fn get_inventory(
&self,
input: &GetInventoryRequest
) -> Result<GetInventoryResult, GetInventoryError>

Query inventory information.

fn get_inventory_schema(
&self,
input: &GetInventorySchemaRequest
) -> Result<GetInventorySchemaResult, GetInventorySchemaError>

Return a list of inventory type names for the account, or return a list of attribute names for a specific Inventory item type.

fn get_maintenance_window(
&self,
input: &GetMaintenanceWindowRequest
) -> Result<GetMaintenanceWindowResult, GetMaintenanceWindowError>

Retrieves a Maintenance Window.

fn get_maintenance_window_execution(
&self,
input: &GetMaintenanceWindowExecutionRequest
) -> Result<GetMaintenanceWindowExecutionResult, GetMaintenanceWindowExecutionError>

Retrieves details about a specific task executed as part of a Maintenance Window execution.

fn get_maintenance_window_execution_task(
&self,
input: &GetMaintenanceWindowExecutionTaskRequest
) -> Result<GetMaintenanceWindowExecutionTaskResult, GetMaintenanceWindowExecutionTaskError>

Retrieves the details about a specific task executed as part of a Maintenance Window execution.

fn get_parameter(
&self,
input: &GetParameterRequest
) -> Result<GetParameterResult, GetParameterError>

Get information about a parameter by using the parameter name.

fn get_parameter_history(
&self,
input: &GetParameterHistoryRequest
) -> Result<GetParameterHistoryResult, GetParameterHistoryError>

Query a list of all parameters used by the AWS account.

fn get_parameters(
&self,
input: &GetParametersRequest
) -> Result<GetParametersResult, GetParametersError>

Get details of a parameter.

fn get_parameters_by_path(
&self,
input: &GetParametersByPathRequest
) -> Result<GetParametersByPathResult, GetParametersByPathError>

Retrieve parameters in a specific hierarchy. For more information, see Using Parameter Hierarchies.

fn get_patch_baseline(
&self,
input: &GetPatchBaselineRequest
) -> Result<GetPatchBaselineResult, GetPatchBaselineError>

Retrieves information about a patch baseline.

fn get_patch_baseline_for_patch_group(
&self,
input: &GetPatchBaselineForPatchGroupRequest
) -> Result<GetPatchBaselineForPatchGroupResult, GetPatchBaselineForPatchGroupError>

Retrieves the patch baseline that should be used for the specified patch group.

fn list_associations(
&self,
input: &ListAssociationsRequest
) -> Result<ListAssociationsResult, ListAssociationsError>

Lists the associations for the specified Systems Manager document or instance.

fn list_command_invocations(
&self,
input: &ListCommandInvocationsRequest
) -> Result<ListCommandInvocationsResult, ListCommandInvocationsError>

An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user executes SendCommand against three instances, then a command invocation is created for each requested instance ID. ListCommandInvocations provide status about command execution.

fn list_commands(
&self,
input: &ListCommandsRequest
) -> Result<ListCommandsResult, ListCommandsError>

Lists the commands requested by users of the AWS account.

fn list_document_versions(
&self,
input: &ListDocumentVersionsRequest
) -> Result<ListDocumentVersionsResult, ListDocumentVersionsError>

List all versions for a document.

fn list_documents(
&self,
input: &ListDocumentsRequest
) -> Result<ListDocumentsResult, ListDocumentsError>

Describes one or more of your SSM documents.

fn list_inventory_entries(
&self,
input: &ListInventoryEntriesRequest
) -> Result<ListInventoryEntriesResult, ListInventoryEntriesError>

A list of inventory items returned by the request.

fn list_tags_for_resource(
&self,
input: &ListTagsForResourceRequest
) -> Result<ListTagsForResourceResult, ListTagsForResourceError>

Returns a list of the tags assigned to the specified resource.

fn modify_document_permission(
&self,
input: &ModifyDocumentPermissionRequest
) -> Result<ModifyDocumentPermissionResponse, ModifyDocumentPermissionError>

Shares a Systems Manager document publicly or privately. If you share a document privately, you must specify the AWS user account IDs for those people who can use the document. If you share a document publicly, you must specify All as the account ID.

fn put_inventory(
&self,
input: &PutInventoryRequest
) -> Result<PutInventoryResult, PutInventoryError>

Bulk update custom inventory items on one more instance. The request adds an inventory item, if it doesn't already exist, or updates an inventory item, if it does exist.

fn put_parameter(
&self,
input: &PutParameterRequest
) -> Result<PutParameterResult, PutParameterError>

Add one or more parameters to the system.

fn register_default_patch_baseline(
&self,
input: &RegisterDefaultPatchBaselineRequest
) -> Result<RegisterDefaultPatchBaselineResult, RegisterDefaultPatchBaselineError>

Defines the default patch baseline.

fn register_patch_baseline_for_patch_group(
&self,
input: &RegisterPatchBaselineForPatchGroupRequest
) -> Result<RegisterPatchBaselineForPatchGroupResult, RegisterPatchBaselineForPatchGroupError>

Registers a patch baseline for a patch group.

fn register_target_with_maintenance_window(
&self,
input: &RegisterTargetWithMaintenanceWindowRequest
) -> Result<RegisterTargetWithMaintenanceWindowResult, RegisterTargetWithMaintenanceWindowError>

Registers a target with a Maintenance Window.

fn register_task_with_maintenance_window(
&self,
input: &RegisterTaskWithMaintenanceWindowRequest
) -> Result<RegisterTaskWithMaintenanceWindowResult, RegisterTaskWithMaintenanceWindowError>

Adds a new task to a Maintenance Window.

fn remove_tags_from_resource(
&self,
input: &RemoveTagsFromResourceRequest
) -> Result<RemoveTagsFromResourceResult, RemoveTagsFromResourceError>

Removes all tags from the specified resource.

fn send_command(
&self,
input: &SendCommandRequest
) -> Result<SendCommandResult, SendCommandError>

Executes commands on one or more managed instances.

fn start_automation_execution(
&self,
input: &StartAutomationExecutionRequest
) -> Result<StartAutomationExecutionResult, StartAutomationExecutionError>

Initiates execution of an Automation document.

fn stop_automation_execution(
&self,
input: &StopAutomationExecutionRequest
) -> Result<StopAutomationExecutionResult, StopAutomationExecutionError>

Stop an Automation that is currently executing.

fn update_association(
&self,
input: &UpdateAssociationRequest
) -> Result<UpdateAssociationResult, UpdateAssociationError>

Updates an association. You can only update the document version, schedule, parameters, and Amazon S3 output of an association.

fn update_association_status(
&self,
input: &UpdateAssociationStatusRequest
) -> Result<UpdateAssociationStatusResult, UpdateAssociationStatusError>

Updates the status of the Systems Manager document associated with the specified instance.

fn update_document(
&self,
input: &UpdateDocumentRequest
) -> Result<UpdateDocumentResult, UpdateDocumentError>

The document you want to update.

fn update_document_default_version(
&self,
input: &UpdateDocumentDefaultVersionRequest
) -> Result<UpdateDocumentDefaultVersionResult, UpdateDocumentDefaultVersionError>

Set the default version of a document.

fn update_maintenance_window(
&self,
input: &UpdateMaintenanceWindowRequest
) -> Result<UpdateMaintenanceWindowResult, UpdateMaintenanceWindowError>

Updates an existing Maintenance Window. Only specified parameters are modified.

fn update_managed_instance_role(
&self,
input: &UpdateManagedInstanceRoleRequest
) -> Result<UpdateManagedInstanceRoleResult, UpdateManagedInstanceRoleError>

Assigns or changes an Amazon Identity and Access Management (IAM) role to the managed instance.

fn update_patch_baseline(
&self,
input: &UpdatePatchBaselineRequest
) -> Result<UpdatePatchBaselineResult, UpdatePatchBaselineError>

Modifies an existing patch baseline. Fields not specified in the request are left unchanged.