Crate rusoto_ssm

Amazon EC2 Systems Manager
Amazon EC2 Systems Manager is a collection of capabilities that helps you automate management tasks such as collecting system inventory, applying operating system (OS) patches, automating the creation of Amazon Machine Images (AMIs), and configuring operating systems (OSs) and applications at scale. Systems Manager lets you remotely and securely manage the configuration of your managed instances. A managed instance is any Amazon EC2 instance or on-premises machine in your hybrid environment that has been configured for Systems Manager.

This reference is intended to be used with the Amazon EC2 Systems Manager User Guide.

To get started, verify prerequisites and configure managed instances. For more information, see Systems Manager Prerequisites.

If you're using the service, you're probably looking for SsmClient and Ssm.

Structs

Activation
An activation registers one or more on-premises servers or virtual machines (VMs) with AWS so that you can configure those servers or VMs using Run Command. A server or VM that has been registered with AWS is called a managed instance.

AddTagsToResourceRequest

AddTagsToResourceResult

Association
Describes an association of a Systems Manager document and an instance.

AssociationDescription
Describes the parameters for a document.

AssociationFilter
Describes a filter.

AssociationOverview
Information about the association.

AssociationStatus
Describes an association status.

AutomationExecution
Detailed information about the current state of an individual Automation execution.

AutomationExecutionFilter
A filter used to match specific automation executions. This is used to limit the scope of Automation execution information returned.

AutomationExecutionMetadata
Details about a specific Automation execution.

CancelCommandRequest

CancelCommandResult
Whether or not the command was successfully canceled. There is no guarantee that a request can be canceled.

Command
Describes a command request.

CommandFilter
Describes a command filter.

CommandInvocation
An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user executes SendCommand against three instances, then a command invocation is created for each requested instance ID. A command invocation returns status and detail information about a command you executed.

CommandPlugin
Describes plugin details.

CreateActivationRequest

CreateActivationResult

CreateAssociationBatchRequest

CreateAssociationBatchRequestEntry
Describes the association of a Systems Manager document and an instance.

CreateAssociationBatchResult

CreateAssociationRequest

CreateAssociationResult

CreateDocumentRequest

CreateDocumentResult

CreateMaintenanceWindowRequest

CreateMaintenanceWindowResult

CreatePatchBaselineRequest

CreatePatchBaselineResult

DeleteActivationRequest

DeleteActivationResult

DeleteAssociationRequest

DeleteAssociationResult

DeleteDocumentRequest

DeleteDocumentResult

DeleteMaintenanceWindowRequest

DeleteMaintenanceWindowResult

DeleteParameterRequest

DeleteParameterResult

DeleteParametersRequest

DeleteParametersResult

DeletePatchBaselineRequest

DeletePatchBaselineResult

DeregisterManagedInstanceRequest

DeregisterManagedInstanceResult

DeregisterPatchBaselineForPatchGroupRequest

DeregisterPatchBaselineForPatchGroupResult

DeregisterTargetFromMaintenanceWindowRequest

DeregisterTargetFromMaintenanceWindowResult

DeregisterTaskFromMaintenanceWindowRequest

DeregisterTaskFromMaintenanceWindowResult

DescribeActivationsFilter
Filter for the DescribeActivation API.

DescribeActivationsRequest

DescribeActivationsResult

DescribeAssociationRequest

DescribeAssociationResult

DescribeAutomationExecutionsRequest

DescribeAutomationExecutionsResult

DescribeAvailablePatchesRequest

DescribeAvailablePatchesResult

DescribeDocumentPermissionRequest

DescribeDocumentPermissionResponse

DescribeDocumentRequest

DescribeDocumentResult

DescribeEffectiveInstanceAssociationsRequest

DescribeEffectiveInstanceAssociationsResult

DescribeEffectivePatchesForPatchBaselineRequest

DescribeEffectivePatchesForPatchBaselineResult

DescribeInstanceAssociationsStatusRequest

DescribeInstanceAssociationsStatusResult

DescribeInstanceInformationRequest

DescribeInstanceInformationResult

DescribeInstancePatchStatesForPatchGroupRequest

DescribeInstancePatchStatesForPatchGroupResult

DescribeInstancePatchStatesRequest

DescribeInstancePatchStatesResult

DescribeInstancePatchesRequest

DescribeInstancePatchesResult

DescribeMaintenanceWindowExecutionTaskInvocationsRequest

DescribeMaintenanceWindowExecutionTaskInvocationsResult

DescribeMaintenanceWindowExecutionTasksRequest

DescribeMaintenanceWindowExecutionTasksResult

DescribeMaintenanceWindowExecutionsRequest

DescribeMaintenanceWindowExecutionsResult

DescribeMaintenanceWindowTargetsRequest

DescribeMaintenanceWindowTargetsResult

DescribeMaintenanceWindowTasksRequest

DescribeMaintenanceWindowTasksResult

DescribeMaintenanceWindowsRequest

DescribeMaintenanceWindowsResult

DescribeParametersRequest

DescribeParametersResult

DescribePatchBaselinesRequest

DescribePatchBaselinesResult

DescribePatchGroupStateRequest

DescribePatchGroupStateResult

DescribePatchGroupsRequest

DescribePatchGroupsResult

DocumentDefaultVersionDescription
A default version of a document.

DocumentDescription
Describes an SSM document.

DocumentFilter
Describes a filter.

DocumentIdentifier
Describes the name of an SSM document.

DocumentParameter
Parameters specified in a System Manager document that execute on the server when the command is run.

DocumentVersionInfo
Version information about the document.

EffectivePatch
The EffectivePatch structure defines metadata about a patch along with the approval state of the patch in a particular patch baseline. The approval state includes information about whether the patch is currently approved, due to be approved by a rule, explicitly approved, or explicitly rejected and the date the patch was or will be approved.

FailedCreateAssociation
Describes a failed association.

FailureDetails
Information about an Automation failure.

GetAutomationExecutionRequest

GetAutomationExecutionResult

GetCommandInvocationRequest

GetCommandInvocationResult

GetDefaultPatchBaselineRequest

GetDefaultPatchBaselineResult

GetDeployablePatchSnapshotForInstanceRequest

GetDeployablePatchSnapshotForInstanceResult

GetDocumentRequest

GetDocumentResult

GetInventoryRequest

GetInventoryResult

GetInventorySchemaRequest

GetInventorySchemaResult

GetMaintenanceWindowExecutionRequest

GetMaintenanceWindowExecutionResult

GetMaintenanceWindowExecutionTaskRequest

GetMaintenanceWindowExecutionTaskResult

GetMaintenanceWindowRequest

GetMaintenanceWindowResult

GetParameterHistoryRequest

GetParameterHistoryResult

GetParameterRequest

GetParameterResult

GetParametersByPathRequest

GetParametersByPathResult

GetParametersRequest

GetParametersResult

GetPatchBaselineForPatchGroupRequest

GetPatchBaselineForPatchGroupResult

GetPatchBaselineRequest

GetPatchBaselineResult

InstanceAggregatedAssociationOverview
Status information about the aggregated associations.

InstanceAssociation
One or more association documents on the instance.

InstanceAssociationOutputLocation
An Amazon S3 bucket where you want to store the results of this request.

InstanceAssociationOutputUrl
The URL of Amazon S3 bucket where you want to store the results of this request.

InstanceAssociationStatusInfo
Status information about the instance association.

InstanceInformation
Describes a filter for a specific list of instances.

InstanceInformationFilter
Describes a filter for a specific list of instances.

InstanceInformationStringFilter
The filters to describe or get information about your managed instances.

InstancePatchState
Defines the high-level patch compliance state for a managed instance, providing information about the number of installed, missing, not applicable, and failed patches along with metadata about the operation when this information was gathered for the instance.

InstancePatchStateFilter
Defines a filter used in DescribeInstancePatchStatesForPatchGroup used to scope down the information returned by the API.

InventoryFilter
One or more filters. Use a filter to return a more specific list of results.

InventoryItem
Information collected from managed instances based on your inventory policy document

InventoryItemAttribute
Attributes are the entries within the inventory item content. It contains name and value.

InventoryItemSchema
The inventory item schema definition. Users can use this to compose inventory query filters.

InventoryResultEntity
Inventory query results.

InventoryResultItem
The inventory result item.

ListAssociationsRequest

ListAssociationsResult

ListCommandInvocationsRequest

ListCommandInvocationsResult

ListCommandsRequest

ListCommandsResult

ListDocumentVersionsRequest

ListDocumentVersionsResult

ListDocumentsRequest

ListDocumentsResult

ListInventoryEntriesRequest

ListInventoryEntriesResult

ListTagsForResourceRequest

ListTagsForResourceResult

LoggingInfo
Information about an Amazon S3 bucket to write instance-level logs to.

MaintenanceWindowExecution
Describes the information about an execution of a Maintenance Window.

MaintenanceWindowExecutionTaskIdentity
Information about a task execution performed as part of a Maintenance Window execution.

MaintenanceWindowExecutionTaskInvocationIdentity
Describes the information about a task invocation for a particular target as part of a task execution performed as part of a Maintenance Window execution.

MaintenanceWindowFilter
Filter used in the request.

MaintenanceWindowIdentity
Information about the Maintenance Window.

MaintenanceWindowTarget
The target registered with the Maintenance Window.

MaintenanceWindowTask
Information about a task defined for a Maintenance Window.

MaintenanceWindowTaskParameterValueExpression
Defines the values for a task parameter.

ModifyDocumentPermissionRequest

ModifyDocumentPermissionResponse

NotificationConfig
Configurations for sending notifications.

Parameter
An Amazon EC2 Systems Manager parameter in Parameter Store.

ParameterHistory
Information about parameter usage.

ParameterMetadata
Metada includes information like the ARN of the last user and the date/time the parameter was last used.

ParameterStringFilter
One or more filters. Use a filter to return a more specific list of results.

ParametersFilter
One or more filters. Use a filter to return a more specific list of results.

Patch
Represents metadata about a patch.

PatchBaselineIdentity
Defines the basic information about a patch baseline.

PatchComplianceData
Information about the state of a patch on a particular instance as it relates to the patch baseline used to patch the instance.

PatchFilter
Defines a patch filter.

PatchFilterGroup
A set of patch filters, typically used for approval rules.

PatchGroupPatchBaselineMapping
The mapping between a patch group and the patch baseline the patch group is registered with.

PatchOrchestratorFilter
Defines a filter used in Patch Manager APIs.

PatchRule
Defines an approval rule for a patch baseline.

PatchRuleGroup
A set of rules defining the approval rules for a patch baseline.

PatchStatus
Information about the approval status of a patch.

PutInventoryRequest

PutInventoryResult

PutParameterRequest

PutParameterResult

RegisterDefaultPatchBaselineRequest

RegisterDefaultPatchBaselineResult

RegisterPatchBaselineForPatchGroupRequest

RegisterPatchBaselineForPatchGroupResult

RegisterTargetWithMaintenanceWindowRequest

RegisterTargetWithMaintenanceWindowResult

RegisterTaskWithMaintenanceWindowRequest

RegisterTaskWithMaintenanceWindowResult

RemoveTagsFromResourceRequest

RemoveTagsFromResourceResult

ResultAttribute
The inventory item result attribute.

S3OutputLocation
An Amazon S3 bucket where you want to store the results of this request.

S3OutputUrl
A URL for the Amazon S3 bucket where you want to store the results of this request.

SendCommandRequest

SendCommandResult

SsmClient
A client for the Amazon SSM API.

StartAutomationExecutionRequest

StartAutomationExecutionResult

StepExecution
Detailed information about an the execution state of an Automation step.

StopAutomationExecutionRequest

StopAutomationExecutionResult

Tag
Metadata that you assign to your managed instances. Tags enable you to categorize your managed instances in different ways, for example, by purpose, owner, or environment.

Target

An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call.

UpdateAssociationRequest

UpdateAssociationResult

UpdateAssociationStatusRequest

UpdateAssociationStatusResult

UpdateDocumentDefaultVersionRequest

UpdateDocumentDefaultVersionResult

UpdateDocumentRequest

UpdateDocumentResult

UpdateMaintenanceWindowRequest

UpdateMaintenanceWindowResult

UpdateManagedInstanceRoleRequest

UpdateManagedInstanceRoleResult

UpdatePatchBaselineRequest

UpdatePatchBaselineResult

Enums

AddTagsToResourceError
Errors returned by AddTagsToResource

CancelCommandError
Errors returned by CancelCommand

CreateActivationError
Errors returned by CreateActivation

CreateAssociationBatchError
Errors returned by CreateAssociationBatch

CreateAssociationError
Errors returned by CreateAssociation

CreateDocumentError
Errors returned by CreateDocument

CreateMaintenanceWindowError
Errors returned by CreateMaintenanceWindow

CreatePatchBaselineError
Errors returned by CreatePatchBaseline

DeleteActivationError
Errors returned by DeleteActivation

DeleteAssociationError
Errors returned by DeleteAssociation

DeleteDocumentError
Errors returned by DeleteDocument

DeleteMaintenanceWindowError
Errors returned by DeleteMaintenanceWindow

DeleteParameterError
Errors returned by DeleteParameter

DeleteParametersError
Errors returned by DeleteParameters

DeletePatchBaselineError
Errors returned by DeletePatchBaseline

DeregisterManagedInstanceError
Errors returned by DeregisterManagedInstance

DeregisterPatchBaselineForPatchGroupError
Errors returned by DeregisterPatchBaselineForPatchGroup

DeregisterTargetFromMaintenanceWindowError
Errors returned by DeregisterTargetFromMaintenanceWindow

DeregisterTaskFromMaintenanceWindowError
Errors returned by DeregisterTaskFromMaintenanceWindow

DescribeActivationsError
Errors returned by DescribeActivations

DescribeAssociationError
Errors returned by DescribeAssociation

DescribeAutomationExecutionsError
Errors returned by DescribeAutomationExecutions

DescribeAvailablePatchesError
Errors returned by DescribeAvailablePatches

DescribeDocumentError
Errors returned by DescribeDocument

DescribeDocumentPermissionError
Errors returned by DescribeDocumentPermission

DescribeEffectiveInstanceAssociationsError
Errors returned by DescribeEffectiveInstanceAssociations

DescribeEffectivePatchesForPatchBaselineError
Errors returned by DescribeEffectivePatchesForPatchBaseline

DescribeInstanceAssociationsStatusError
Errors returned by DescribeInstanceAssociationsStatus

DescribeInstanceInformationError
Errors returned by DescribeInstanceInformation

DescribeInstancePatchStatesError
Errors returned by DescribeInstancePatchStates

DescribeInstancePatchStatesForPatchGroupError
Errors returned by DescribeInstancePatchStatesForPatchGroup

DescribeInstancePatchesError
Errors returned by DescribeInstancePatches

DescribeMaintenanceWindowExecutionTaskInvocationsError
Errors returned by DescribeMaintenanceWindowExecutionTaskInvocations

DescribeMaintenanceWindowExecutionTasksError
Errors returned by DescribeMaintenanceWindowExecutionTasks

DescribeMaintenanceWindowExecutionsError
Errors returned by DescribeMaintenanceWindowExecutions

DescribeMaintenanceWindowTargetsError
Errors returned by DescribeMaintenanceWindowTargets

DescribeMaintenanceWindowTasksError
Errors returned by DescribeMaintenanceWindowTasks

DescribeMaintenanceWindowsError
Errors returned by DescribeMaintenanceWindows

DescribeParametersError
Errors returned by DescribeParameters

DescribePatchBaselinesError
Errors returned by DescribePatchBaselines

DescribePatchGroupStateError
Errors returned by DescribePatchGroupState

DescribePatchGroupsError
Errors returned by DescribePatchGroups

GetAutomationExecutionError
Errors returned by GetAutomationExecution

GetCommandInvocationError
Errors returned by GetCommandInvocation

GetDefaultPatchBaselineError
Errors returned by GetDefaultPatchBaseline

GetDeployablePatchSnapshotForInstanceError
Errors returned by GetDeployablePatchSnapshotForInstance

GetDocumentError
Errors returned by GetDocument

GetInventoryError
Errors returned by GetInventory

GetInventorySchemaError
Errors returned by GetInventorySchema

GetMaintenanceWindowError
Errors returned by GetMaintenanceWindow

GetMaintenanceWindowExecutionError
Errors returned by GetMaintenanceWindowExecution

GetMaintenanceWindowExecutionTaskError
Errors returned by GetMaintenanceWindowExecutionTask

GetParameterError
Errors returned by GetParameter

GetParameterHistoryError
Errors returned by GetParameterHistory

GetParametersByPathError
Errors returned by GetParametersByPath

GetParametersError
Errors returned by GetParameters

GetPatchBaselineError
Errors returned by GetPatchBaseline

GetPatchBaselineForPatchGroupError
Errors returned by GetPatchBaselineForPatchGroup

ListAssociationsError
Errors returned by ListAssociations

ListCommandInvocationsError
Errors returned by ListCommandInvocations

ListCommandsError
Errors returned by ListCommands

ListDocumentVersionsError
Errors returned by ListDocumentVersions

ListDocumentsError
Errors returned by ListDocuments

ListInventoryEntriesError
Errors returned by ListInventoryEntries

ListTagsForResourceError
Errors returned by ListTagsForResource

ModifyDocumentPermissionError
Errors returned by ModifyDocumentPermission

PutInventoryError
Errors returned by PutInventory

PutParameterError
Errors returned by PutParameter

RegisterDefaultPatchBaselineError
Errors returned by RegisterDefaultPatchBaseline

RegisterPatchBaselineForPatchGroupError
Errors returned by RegisterPatchBaselineForPatchGroup

RegisterTargetWithMaintenanceWindowError
Errors returned by RegisterTargetWithMaintenanceWindow

RegisterTaskWithMaintenanceWindowError
Errors returned by RegisterTaskWithMaintenanceWindow

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

SendCommandError
Errors returned by SendCommand

StartAutomationExecutionError
Errors returned by StartAutomationExecution

StopAutomationExecutionError
Errors returned by StopAutomationExecution

UpdateAssociationError
Errors returned by UpdateAssociation

UpdateAssociationStatusError
Errors returned by UpdateAssociationStatus

UpdateDocumentDefaultVersionError
Errors returned by UpdateDocumentDefaultVersion

UpdateDocumentError
Errors returned by UpdateDocument

UpdateMaintenanceWindowError
Errors returned by UpdateMaintenanceWindow

UpdateManagedInstanceRoleError
Errors returned by UpdateManagedInstanceRole

UpdatePatchBaselineError
Errors returned by UpdatePatchBaseline

Traits

Ssm
Trait representing the capabilities of the Amazon SSM API. Amazon SSM clients implement this trait.