Crate rusoto_ssm

AWS Systems Manager
AWS Systems Manager is a collection of capabilities that helps you automate management tasks such as collecting system inventory, applying operating system (OS) patches, automating the creation of Amazon Machine Images (AMIs), and configuring operating systems (OSs) and applications at scale. Systems Manager lets you remotely and securely manage the configuration of your managed instances. A managed instance is any Amazon EC2 instance or on-premises machine in your hybrid environment that has been configured for Systems Manager.

This reference is intended to be used with the AWS Systems Manager User Guide.

To get started, verify prerequisites and configure managed instances. For more information, see Systems Manager Prerequisites in the AWS Systems Manager User Guide.

For information about other API actions you can perform on Amazon EC2 instances, see the Amazon EC2 API Reference. For information about how to use a Query API, see Making API Requests.

If you're using the service, you're probably looking for SsmClient and Ssm.

Structs

Activation
An activation registers one or more on-premises servers or virtual machines (VMs) with AWS so that you can configure those servers or VMs using Run Command. A server or VM that has been registered with AWS is called a managed instance.

AddTagsToResourceRequest

AddTagsToResourceResult

Association
Describes an association of a Systems Manager document and an instance.

AssociationDescription
Describes the parameters for a document.

AssociationExecution
Includes information about the specified association.

AssociationExecutionFilter
Filters used in the request.

AssociationExecutionTarget
Includes information about the specified association execution.

AssociationExecutionTargetsFilter
Filters for the association execution.

AssociationFilter
Describes a filter.

AssociationOverview
Information about the association.

AssociationStatus
Describes an association status.

AssociationVersionInfo
Information about the association version.

AutomationExecution
Detailed information about the current state of an individual Automation execution.

AutomationExecutionFilter
A filter used to match specific automation executions. This is used to limit the scope of Automation execution information returned.

AutomationExecutionMetadata
Details about a specific Automation execution.

CancelCommandRequest

CancelCommandResult
Whether or not the command was successfully canceled. There is no guarantee that a request can be canceled.

CloudWatchOutputConfig
Configuration options for sending command output to CloudWatch Logs.

Command
Describes a command request.

CommandFilter
Describes a command filter.

CommandInvocation
An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user executes SendCommand against three instances, then a command invocation is created for each requested instance ID. A command invocation returns status and detail information about a command you executed.

CommandPlugin
Describes plugin details.

ComplianceExecutionSummary
A summary of the call execution that includes an execution ID, the type of execution (for example, Command), and the date/time of the execution using a datetime object that is saved in the following format: yyyy-MM-dd'T'HH:mm:ss'Z'.

ComplianceItem
Information about the compliance as defined by the resource type. For example, for a patch resource type, Items includes information about the PatchSeverity, Classification, etc.

ComplianceItemEntry
Information about a compliance item.

ComplianceStringFilter
One or more filters. Use a filter to return a more specific list of results.

ComplianceSummaryItem
A summary of compliance information by compliance type.

CompliantSummary
A summary of resources that are compliant. The summary is organized according to the resource count for each compliance type.

CreateActivationRequest

CreateActivationResult

CreateAssociationBatchRequest

CreateAssociationBatchRequestEntry
Describes the association of a Systems Manager document and an instance.

CreateAssociationBatchResult

CreateAssociationRequest

CreateAssociationResult

CreateDocumentRequest

CreateDocumentResult

CreateMaintenanceWindowRequest

CreateMaintenanceWindowResult

CreatePatchBaselineRequest

CreatePatchBaselineResult

CreateResourceDataSyncRequest

CreateResourceDataSyncResult

DeleteActivationRequest

DeleteActivationResult

DeleteAssociationRequest

DeleteAssociationResult

DeleteDocumentRequest

DeleteDocumentResult

DeleteInventoryRequest

DeleteInventoryResult

DeleteMaintenanceWindowRequest

DeleteMaintenanceWindowResult

DeleteParameterRequest

DeleteParameterResult

DeleteParametersRequest

DeleteParametersResult

DeletePatchBaselineRequest

DeletePatchBaselineResult

DeleteResourceDataSyncRequest

DeleteResourceDataSyncResult

DeregisterManagedInstanceRequest

DeregisterManagedInstanceResult

DeregisterPatchBaselineForPatchGroupRequest

DeregisterPatchBaselineForPatchGroupResult

DeregisterTargetFromMaintenanceWindowRequest

DeregisterTargetFromMaintenanceWindowResult

DeregisterTaskFromMaintenanceWindowRequest

DeregisterTaskFromMaintenanceWindowResult

DescribeActivationsFilter
Filter for the DescribeActivation API.

DescribeActivationsRequest

DescribeActivationsResult

DescribeAssociationExecutionTargetsRequest

DescribeAssociationExecutionTargetsResult

DescribeAssociationExecutionsRequest

DescribeAssociationExecutionsResult

DescribeAssociationRequest

DescribeAssociationResult

DescribeAutomationExecutionsRequest

DescribeAutomationExecutionsResult

DescribeAutomationStepExecutionsRequest

DescribeAutomationStepExecutionsResult

DescribeAvailablePatchesRequest

DescribeAvailablePatchesResult

DescribeDocumentPermissionRequest

DescribeDocumentPermissionResponse

DescribeDocumentRequest

DescribeDocumentResult

DescribeEffectiveInstanceAssociationsRequest

DescribeEffectiveInstanceAssociationsResult

DescribeEffectivePatchesForPatchBaselineRequest

DescribeEffectivePatchesForPatchBaselineResult

DescribeInstanceAssociationsStatusRequest

DescribeInstanceAssociationsStatusResult

DescribeInstanceInformationRequest

DescribeInstanceInformationResult

DescribeInstancePatchStatesForPatchGroupRequest

DescribeInstancePatchStatesForPatchGroupResult

DescribeInstancePatchStatesRequest

DescribeInstancePatchStatesResult

DescribeInstancePatchesRequest

DescribeInstancePatchesResult

DescribeInventoryDeletionsRequest

DescribeInventoryDeletionsResult

DescribeMaintenanceWindowExecutionTaskInvocationsRequest

DescribeMaintenanceWindowExecutionTaskInvocationsResult

DescribeMaintenanceWindowExecutionTasksRequest

DescribeMaintenanceWindowExecutionTasksResult

DescribeMaintenanceWindowExecutionsRequest

DescribeMaintenanceWindowExecutionsResult

DescribeMaintenanceWindowTargetsRequest

DescribeMaintenanceWindowTargetsResult

DescribeMaintenanceWindowTasksRequest

DescribeMaintenanceWindowTasksResult

DescribeMaintenanceWindowsRequest

DescribeMaintenanceWindowsResult

DescribeParametersRequest

DescribeParametersResult

DescribePatchBaselinesRequest

DescribePatchBaselinesResult

DescribePatchGroupStateRequest

DescribePatchGroupStateResult

DescribePatchGroupsRequest

DescribePatchGroupsResult

DocumentDefaultVersionDescription
A default version of a document.

DocumentDescription
Describes a Systems Manager document.

DocumentFilter
Describes a filter.

DocumentIdentifier
Describes the name of a Systems Manager document.

DocumentKeyValuesFilter
One or more filters. Use a filter to return a more specific list of documents.

For keys, you can specify one or more tags that have been applied to a document.

Other valid values include Owner, Name, PlatformTypes, and DocumentType.

Note that only one Owner can be specified in a request. For example: Key=Owner,Values=Self.

If you use Name as a key, you can use a name prefix to return a list of documents. For example, in the AWS CLI, to return a list of all documents that begin with Te, run the following command:

aws ssm list-documents --filters Key=Name,Values=Te

If you specify more than two keys, only documents that are identified by all the tags are returned in the results. If you specify more than two values for a key, documents that are identified by any of the values are returned in the results.

To specify a custom key and value pair, use the format Key=tag:[tagName],Values=[valueName].

For example, if you created a Key called region and are using the AWS CLI to call the list-documents command:

aws ssm list-documents --filters Key=tag:region,Values=east,west Key=Owner,Values=Self

DocumentParameter
Parameters specified in a System Manager document that execute on the server when the command is run.

DocumentVersionInfo
Version information about the document.

EffectivePatch
The EffectivePatch structure defines metadata about a patch along with the approval state of the patch in a particular patch baseline. The approval state includes information about whether the patch is currently approved, due to be approved by a rule, explicitly approved, or explicitly rejected and the date the patch was or will be approved.

FailedCreateAssociation
Describes a failed association.

FailureDetails
Information about an Automation failure.

GetAutomationExecutionRequest

GetAutomationExecutionResult

GetCommandInvocationRequest

GetCommandInvocationResult

GetDefaultPatchBaselineRequest

GetDefaultPatchBaselineResult

GetDeployablePatchSnapshotForInstanceRequest

GetDeployablePatchSnapshotForInstanceResult

GetDocumentRequest

GetDocumentResult

GetInventoryRequest

GetInventoryResult

GetInventorySchemaRequest

GetInventorySchemaResult

GetMaintenanceWindowExecutionRequest

GetMaintenanceWindowExecutionResult

GetMaintenanceWindowExecutionTaskInvocationRequest

GetMaintenanceWindowExecutionTaskInvocationResult

GetMaintenanceWindowExecutionTaskRequest

GetMaintenanceWindowExecutionTaskResult

GetMaintenanceWindowRequest

GetMaintenanceWindowResult

GetMaintenanceWindowTaskRequest

GetMaintenanceWindowTaskResult

GetParameterHistoryRequest

GetParameterHistoryResult

GetParameterRequest

GetParameterResult

GetParametersByPathRequest

GetParametersByPathResult

GetParametersRequest

GetParametersResult

GetPatchBaselineForPatchGroupRequest

GetPatchBaselineForPatchGroupResult

GetPatchBaselineRequest

GetPatchBaselineResult

InstanceAggregatedAssociationOverview
Status information about the aggregated associations.

InstanceAssociation
One or more association documents on the instance.

InstanceAssociationOutputLocation
An Amazon S3 bucket where you want to store the results of this request.

InstanceAssociationOutputUrl
The URL of Amazon S3 bucket where you want to store the results of this request.

InstanceAssociationStatusInfo
Status information about the instance association.

InstanceInformation
Describes a filter for a specific list of instances.

InstanceInformationFilter
Describes a filter for a specific list of instances.

InstanceInformationStringFilter
The filters to describe or get information about your managed instances.

InstancePatchState
Defines the high-level patch compliance state for a managed instance, providing information about the number of installed, missing, not applicable, and failed patches along with metadata about the operation when this information was gathered for the instance.

InstancePatchStateFilter
Defines a filter used in DescribeInstancePatchStatesForPatchGroup used to scope down the information returned by the API.

InventoryAggregator
Specifies the inventory type and attribute for the aggregation execution.

InventoryDeletionStatusItem
Status information returned by the DeleteInventory action.

InventoryDeletionSummary
Information about the delete operation.

InventoryDeletionSummaryItem
Either a count, remaining count, or a version number in a delete inventory summary.

InventoryFilter
One or more filters. Use a filter to return a more specific list of results.

InventoryItem
Information collected from managed instances based on your inventory policy document

InventoryItemAttribute
Attributes are the entries within the inventory item content. It contains name and value.

InventoryItemSchema
The inventory item schema definition. Users can use this to compose inventory query filters.

InventoryResultEntity
Inventory query results.

InventoryResultItem
The inventory result item.

LabelParameterVersionRequest

LabelParameterVersionResult

ListAssociationVersionsRequest

ListAssociationVersionsResult

ListAssociationsRequest

ListAssociationsResult

ListCommandInvocationsRequest

ListCommandInvocationsResult

ListCommandsRequest

ListCommandsResult

ListComplianceItemsRequest

ListComplianceItemsResult

ListComplianceSummariesRequest

ListComplianceSummariesResult

ListDocumentVersionsRequest

ListDocumentVersionsResult

ListDocumentsRequest

ListDocumentsResult

ListInventoryEntriesRequest

ListInventoryEntriesResult

ListResourceComplianceSummariesRequest

ListResourceComplianceSummariesResult

ListResourceDataSyncRequest

ListResourceDataSyncResult

ListTagsForResourceRequest

ListTagsForResourceResult

LoggingInfo

Information about an Amazon S3 bucket to write instance-level logs to.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported Maintenance Window task types, see MaintenanceWindowTaskInvocationParameters.

MaintenanceWindowAutomationParameters
The parameters for an AUTOMATION task type.

MaintenanceWindowExecution
Describes the information about an execution of a Maintenance Window.

MaintenanceWindowExecutionTaskIdentity
Information about a task execution performed as part of a Maintenance Window execution.

MaintenanceWindowExecutionTaskInvocationIdentity
Describes the information about a task invocation for a particular target as part of a task execution performed as part of a Maintenance Window execution.

MaintenanceWindowFilter
Filter used in the request.

MaintenanceWindowIdentity
Information about the Maintenance Window.

MaintenanceWindowLambdaParameters

The parameters for a LAMBDA task type.

For information about specifying and updating task parameters, see RegisterTaskWithMaintenanceWindow and UpdateMaintenanceWindowTask.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported Maintenance Window task types, see MaintenanceWindowTaskInvocationParameters.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported Maintenance Window task types, see MaintenanceWindowTaskInvocationParameters.

For Lambda tasks, Systems Manager ignores any values specified for TaskParameters and LoggingInfo.

MaintenanceWindowRunCommandParameters

The parameters for a RUN_COMMAND task type.

For information about specifying and updating task parameters, see RegisterTaskWithMaintenanceWindow and UpdateMaintenanceWindowTask.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported Maintenance Window task types, see MaintenanceWindowTaskInvocationParameters.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported Maintenance Window task types, see MaintenanceWindowTaskInvocationParameters.

For Run Command tasks, Systems Manager uses specified values for TaskParameters and LoggingInfo only if no values are specified for TaskInvocationParameters.

MaintenanceWindowStepFunctionsParameters

The parameters for a STEP_FUNCTION task.

For information about specifying and updating task parameters, see RegisterTaskWithMaintenanceWindow and UpdateMaintenanceWindowTask.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported Maintenance Window task types, see MaintenanceWindowTaskInvocationParameters.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported Maintenance Window task types, see MaintenanceWindowTaskInvocationParameters.

For Step Functions tasks, Systems Manager ignores any values specified for TaskParameters and LoggingInfo.

MaintenanceWindowTarget
The target registered with the Maintenance Window.

MaintenanceWindowTask
Information about a task defined for a Maintenance Window.

MaintenanceWindowTaskInvocationParameters
The parameters for task execution.

MaintenanceWindowTaskParameterValueExpression
Defines the values for a task parameter.

ModifyDocumentPermissionRequest

ModifyDocumentPermissionResponse

NonCompliantSummary
A summary of resources that are not compliant. The summary is organized according to resource type.

NotificationConfig
Configurations for sending notifications.

OutputSource
Information about the source where the association execution details are stored.

Parameter
An Amazon EC2 Systems Manager parameter in Parameter Store.

ParameterHistory
Information about parameter usage.

ParameterMetadata
Metada includes information like the ARN of the last user and the date/time the parameter was last used.

ParameterStringFilter

One or more filters. Use a filter to return a more specific list of results.

The Name field can't be used with the GetParametersByPath API action.

ParametersFilter
This data type is deprecated. Instead, use ParameterStringFilter.

Patch
Represents metadata about a patch.

PatchBaselineIdentity
Defines the basic information about a patch baseline.

PatchComplianceData
Information about the state of a patch on a particular instance as it relates to the patch baseline used to patch the instance.

PatchFilter

Defines a patch filter.

A patch filter consists of key/value pairs, but not all keys are valid for all operating system types. For example, the key PRODUCT is valid for all supported operating system types. The key MSRCSEVERITY, however, is valid only for Windows operating systems, and the key SECTION is valid only for Ubuntu operating systems.

Refer to the following sections for information about which keys may be used with each major operating system, and which values are valid for each key.

Windows Operating Systems

The supported keys for Windows operating systems are PRODUCT, CLASSIFICATION, and MSRCSEVERITY. See the following lists for valid values for each of these keys.

Supported key: PRODUCT

Supported values:

Windows7

Windows8

Windows8.1

Windows8Embedded

Windows10

Windows10LTSB

WindowsServer2008

WindowsServer2008R2

WindowsServer2012

WindowsServer2012R2

WindowsServer2016

Use a wildcard character () to target all supported operating system versions.

Supported key: CLASSIFICATION

Supported values:

CriticalUpdates

DefinitionUpdates

Drivers

FeaturePacks

SecurityUpdates

ServicePacks

Tools

UpdateRollups

Updates

Upgrades

Supported key: MSRC_SEVERITY

Supported values:

Critical

Important

Moderate

Low

Unspecified

Ubuntu Operating Systems

The supported keys for Ubuntu operating systems are PRODUCT, PRIORITY, and SECTION. See the following lists for valid values for each of these keys.

Supported key: PRODUCT

Supported values:

Ubuntu14.04

Ubuntu16.04

Use a wildcard character () to target all supported operating system versions.

Supported key: PRIORITY

Supported values:

Required

Important

Standard

Optional

Extra

Supported key: SECTION

Only the length of the key value is validated. Minimum length is 1. Maximum length is 64.

Amazon Linux Operating Systems

The supported keys for Amazon Linux operating systems are PRODUCT, CLASSIFICATION, and SEVERITY. See the following lists for valid values for each of these keys.

Supported key: PRODUCT

Supported values:

AmazonLinux2012.03

AmazonLinux2012.09

AmazonLinux2013.03

AmazonLinux2013.09

AmazonLinux2014.03

AmazonLinux2014.09

AmazonLinux2015.03

AmazonLinux2015.09

AmazonLinux2016.03

AmazonLinux2016.09

AmazonLinux2017.03

AmazonLinux2017.09

Use a wildcard character () to target all supported operating system versions.

Supported key: CLASSIFICATION

Supported values:

Security

Bugfix

Enhancement

Recommended

Newpackage

Supported key: SEVERITY

Supported values:

Critical

Important

Medium

Low

Amazon Linux 2 Operating Systems

The supported keys for Amazon Linux 2 operating systems are PRODUCT, CLASSIFICATION, and SEVERITY. See the following lists for valid values for each of these keys.

Supported key: PRODUCT

Supported values:

AmazonLinux2

AmazonLinux2.0

Use a wildcard character () to target all supported operating system versions.

Supported key: CLASSIFICATION

Supported values:

Security

Bugfix

Enhancement

Recommended

Newpackage

Supported key: SEVERITY

Supported values:

Critical

Important

Medium

Low

RedHat Enterprise Linux (RHEL) Operating Systems

The supported keys for RedHat Enterprise Linux operating systems are PRODUCT, CLASSIFICATION, and SEVERITY. See the following lists for valid values for each of these keys.

Supported key: PRODUCT

Supported values:

RedhatEnterpriseLinux6.5

RedhatEnterpriseLinux6.6

RedhatEnterpriseLinux6.7

RedhatEnterpriseLinux6.8

RedhatEnterpriseLinux6.9

RedhatEnterpriseLinux7.0

RedhatEnterpriseLinux7.1

RedhatEnterpriseLinux7.2

RedhatEnterpriseLinux7.3

RedhatEnterpriseLinux7.4

Use a wildcard character () to target all supported operating system versions.

Supported key: CLASSIFICATION

Supported values:

Security

Bugfix

Enhancement

Recommended

Newpackage

Supported key: SEVERITY

Supported values:

Critical

Important

Medium

Low

SUSE Linux Enterprise Server (SLES) Operating Systems

The supported keys for SLES operating systems are PRODUCT, CLASSIFICATION, and SEVERITY. See the following lists for valid values for each of these keys.

Supported key: PRODUCT

Supported values:

Suse12.0

Suse12.1

Suse12.2

Suse12.3

Suse12.4

Suse12.5

Suse12.6

Suse12.7

Suse12.8

Suse12.9

Use a wildcard character () to target all supported operating system versions.

Supported key: CLASSIFICATION

Supported values:

Security

Recommended

Optional

Feature

Document

Yast

Supported key: SEVERITY

Supported values:

Critical

Important

Moderate

Low

CentOS Operating Systems

The supported keys for CentOS operating systems are PRODUCT, CLASSIFICATION, and SEVERITY. See the following lists for valid values for each of these keys.

Supported key: PRODUCT

Supported values:

CentOS6.5

CentOS6.6

CentOS6.7

CentOS6.8

CentOS6.9

CentOS7.0

CentOS7.1

CentOS7.2

CentOS7.3

CentOS7.4

Use a wildcard character () to target all supported operating system versions.

Supported key: CLASSIFICATION

Supported values:

Security

Bugfix

Enhancement

Recommended

Newpackage

Supported key: SEVERITY

Supported values:

Critical

Important

Medium

Low

PatchFilterGroup
A set of patch filters, typically used for approval rules.

PatchGroupPatchBaselineMapping
The mapping between a patch group and the patch baseline the patch group is registered with.

PatchOrchestratorFilter
Defines a filter used in Patch Manager APIs.

PatchRule
Defines an approval rule for a patch baseline.

PatchRuleGroup
A set of rules defining the approval rules for a patch baseline.

PatchSource
Information about the patches to use to update the instances, including target operating systems and source repository. Applies to Linux instances only.

PatchStatus
Information about the approval status of a patch.

PutComplianceItemsRequest

PutComplianceItemsResult

PutInventoryRequest

PutInventoryResult

PutParameterRequest

PutParameterResult

RegisterDefaultPatchBaselineRequest

RegisterDefaultPatchBaselineResult

RegisterPatchBaselineForPatchGroupRequest

RegisterPatchBaselineForPatchGroupResult

RegisterTargetWithMaintenanceWindowRequest

RegisterTargetWithMaintenanceWindowResult

RegisterTaskWithMaintenanceWindowRequest

RegisterTaskWithMaintenanceWindowResult

RemoveTagsFromResourceRequest

RemoveTagsFromResourceResult

ResolvedTargets
Information about targets that resolved during the Automation execution.

ResourceComplianceSummaryItem
Compliance summary information for a specific resource.

ResourceDataSyncItem
Information about a Resource Data Sync configuration, including its current status and last successful sync.

ResourceDataSyncS3Destination
Information about the target Amazon S3 bucket for the Resource Data Sync.

ResultAttribute
The inventory item result attribute.

S3OutputLocation
An Amazon S3 bucket where you want to store the results of this request.

S3OutputUrl
A URL for the Amazon S3 bucket where you want to store the results of this request.

SendAutomationSignalRequest

SendAutomationSignalResult

SendCommandRequest

SendCommandResult

SeveritySummary
The number of managed instances found for each patch severity level defined in the request filter.

SsmClient
A client for the Amazon SSM API.

StartAssociationsOnceRequest

StartAssociationsOnceResult

StartAutomationExecutionRequest

StartAutomationExecutionResult

StepExecution
Detailed information about an the execution state of an Automation step.

StepExecutionFilter
A filter to limit the amount of step execution information returned by the call.

StopAutomationExecutionRequest

StopAutomationExecutionResult

Tag
Metadata that you assign to your AWS resources. Tags enable you to categorize your resources in different ways, for example, by purpose, owner, or environment. In Systems Manager, you can apply tags to documents, managed instances, Maintenance Windows, Parameter Store parameters, and patch baselines.

Target

An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call.

UpdateAssociationRequest

UpdateAssociationResult

UpdateAssociationStatusRequest

UpdateAssociationStatusResult

UpdateDocumentDefaultVersionRequest

UpdateDocumentDefaultVersionResult

UpdateDocumentRequest

UpdateDocumentResult

UpdateMaintenanceWindowRequest

UpdateMaintenanceWindowResult

UpdateMaintenanceWindowTargetRequest

UpdateMaintenanceWindowTargetResult

UpdateMaintenanceWindowTaskRequest

UpdateMaintenanceWindowTaskResult

UpdateManagedInstanceRoleRequest

UpdateManagedInstanceRoleResult

UpdatePatchBaselineRequest

UpdatePatchBaselineResult

Enums

AddTagsToResourceError
Errors returned by AddTagsToResource

CancelCommandError
Errors returned by CancelCommand

CreateActivationError
Errors returned by CreateActivation

CreateAssociationBatchError
Errors returned by CreateAssociationBatch

CreateAssociationError
Errors returned by CreateAssociation

CreateDocumentError
Errors returned by CreateDocument

CreateMaintenanceWindowError
Errors returned by CreateMaintenanceWindow

CreatePatchBaselineError
Errors returned by CreatePatchBaseline

CreateResourceDataSyncError
Errors returned by CreateResourceDataSync

DeleteActivationError
Errors returned by DeleteActivation

DeleteAssociationError
Errors returned by DeleteAssociation

DeleteDocumentError
Errors returned by DeleteDocument

DeleteInventoryError
Errors returned by DeleteInventory

DeleteMaintenanceWindowError
Errors returned by DeleteMaintenanceWindow

DeleteParameterError
Errors returned by DeleteParameter

DeleteParametersError
Errors returned by DeleteParameters

DeletePatchBaselineError
Errors returned by DeletePatchBaseline

DeleteResourceDataSyncError
Errors returned by DeleteResourceDataSync

DeregisterManagedInstanceError
Errors returned by DeregisterManagedInstance

DeregisterPatchBaselineForPatchGroupError
Errors returned by DeregisterPatchBaselineForPatchGroup

DeregisterTargetFromMaintenanceWindowError
Errors returned by DeregisterTargetFromMaintenanceWindow

DeregisterTaskFromMaintenanceWindowError
Errors returned by DeregisterTaskFromMaintenanceWindow

DescribeActivationsError
Errors returned by DescribeActivations

DescribeAssociationError
Errors returned by DescribeAssociation

DescribeAssociationExecutionTargetsError
Errors returned by DescribeAssociationExecutionTargets

DescribeAssociationExecutionsError
Errors returned by DescribeAssociationExecutions

DescribeAutomationExecutionsError
Errors returned by DescribeAutomationExecutions

DescribeAutomationStepExecutionsError
Errors returned by DescribeAutomationStepExecutions

DescribeAvailablePatchesError
Errors returned by DescribeAvailablePatches

DescribeDocumentError
Errors returned by DescribeDocument

DescribeDocumentPermissionError
Errors returned by DescribeDocumentPermission

DescribeEffectiveInstanceAssociationsError
Errors returned by DescribeEffectiveInstanceAssociations

DescribeEffectivePatchesForPatchBaselineError
Errors returned by DescribeEffectivePatchesForPatchBaseline

DescribeInstanceAssociationsStatusError
Errors returned by DescribeInstanceAssociationsStatus

DescribeInstanceInformationError
Errors returned by DescribeInstanceInformation

DescribeInstancePatchStatesError
Errors returned by DescribeInstancePatchStates

DescribeInstancePatchStatesForPatchGroupError
Errors returned by DescribeInstancePatchStatesForPatchGroup

DescribeInstancePatchesError
Errors returned by DescribeInstancePatches

DescribeInventoryDeletionsError
Errors returned by DescribeInventoryDeletions

DescribeMaintenanceWindowExecutionTaskInvocationsError
Errors returned by DescribeMaintenanceWindowExecutionTaskInvocations

DescribeMaintenanceWindowExecutionTasksError
Errors returned by DescribeMaintenanceWindowExecutionTasks

DescribeMaintenanceWindowExecutionsError
Errors returned by DescribeMaintenanceWindowExecutions

DescribeMaintenanceWindowTargetsError
Errors returned by DescribeMaintenanceWindowTargets

DescribeMaintenanceWindowTasksError
Errors returned by DescribeMaintenanceWindowTasks

DescribeMaintenanceWindowsError
Errors returned by DescribeMaintenanceWindows

DescribeParametersError
Errors returned by DescribeParameters

DescribePatchBaselinesError
Errors returned by DescribePatchBaselines

DescribePatchGroupStateError
Errors returned by DescribePatchGroupState

DescribePatchGroupsError
Errors returned by DescribePatchGroups

GetAutomationExecutionError
Errors returned by GetAutomationExecution

GetCommandInvocationError
Errors returned by GetCommandInvocation

GetDefaultPatchBaselineError
Errors returned by GetDefaultPatchBaseline

GetDeployablePatchSnapshotForInstanceError
Errors returned by GetDeployablePatchSnapshotForInstance

GetDocumentError
Errors returned by GetDocument

GetInventoryError
Errors returned by GetInventory

GetInventorySchemaError
Errors returned by GetInventorySchema

GetMaintenanceWindowError
Errors returned by GetMaintenanceWindow

GetMaintenanceWindowExecutionError
Errors returned by GetMaintenanceWindowExecution

GetMaintenanceWindowExecutionTaskError
Errors returned by GetMaintenanceWindowExecutionTask

GetMaintenanceWindowExecutionTaskInvocationError
Errors returned by GetMaintenanceWindowExecutionTaskInvocation

GetMaintenanceWindowTaskError
Errors returned by GetMaintenanceWindowTask

GetParameterError
Errors returned by GetParameter

GetParameterHistoryError
Errors returned by GetParameterHistory

GetParametersByPathError
Errors returned by GetParametersByPath

GetParametersError
Errors returned by GetParameters

GetPatchBaselineError
Errors returned by GetPatchBaseline

GetPatchBaselineForPatchGroupError
Errors returned by GetPatchBaselineForPatchGroup

LabelParameterVersionError
Errors returned by LabelParameterVersion

ListAssociationVersionsError
Errors returned by ListAssociationVersions

ListAssociationsError
Errors returned by ListAssociations

ListCommandInvocationsError
Errors returned by ListCommandInvocations

ListCommandsError
Errors returned by ListCommands

ListComplianceItemsError
Errors returned by ListComplianceItems

ListComplianceSummariesError
Errors returned by ListComplianceSummaries

ListDocumentVersionsError
Errors returned by ListDocumentVersions

ListDocumentsError
Errors returned by ListDocuments

ListInventoryEntriesError
Errors returned by ListInventoryEntries

ListResourceComplianceSummariesError
Errors returned by ListResourceComplianceSummaries

ListResourceDataSyncError
Errors returned by ListResourceDataSync

ListTagsForResourceError
Errors returned by ListTagsForResource

ModifyDocumentPermissionError
Errors returned by ModifyDocumentPermission

PutComplianceItemsError
Errors returned by PutComplianceItems

PutInventoryError
Errors returned by PutInventory

PutParameterError
Errors returned by PutParameter

RegisterDefaultPatchBaselineError
Errors returned by RegisterDefaultPatchBaseline

RegisterPatchBaselineForPatchGroupError
Errors returned by RegisterPatchBaselineForPatchGroup

RegisterTargetWithMaintenanceWindowError
Errors returned by RegisterTargetWithMaintenanceWindow

RegisterTaskWithMaintenanceWindowError
Errors returned by RegisterTaskWithMaintenanceWindow

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

SendAutomationSignalError
Errors returned by SendAutomationSignal

SendCommandError
Errors returned by SendCommand

StartAssociationsOnceError
Errors returned by StartAssociationsOnce

StartAutomationExecutionError
Errors returned by StartAutomationExecution

StopAutomationExecutionError
Errors returned by StopAutomationExecution

UpdateAssociationError
Errors returned by UpdateAssociation

UpdateAssociationStatusError
Errors returned by UpdateAssociationStatus

UpdateDocumentDefaultVersionError
Errors returned by UpdateDocumentDefaultVersion

UpdateDocumentError
Errors returned by UpdateDocument

UpdateMaintenanceWindowError
Errors returned by UpdateMaintenanceWindow

UpdateMaintenanceWindowTargetError
Errors returned by UpdateMaintenanceWindowTarget

UpdateMaintenanceWindowTaskError
Errors returned by UpdateMaintenanceWindowTask

UpdateManagedInstanceRoleError
Errors returned by UpdateManagedInstanceRole

UpdatePatchBaselineError
Errors returned by UpdatePatchBaseline

Traits

Ssm
Trait representing the capabilities of the Amazon SSM API. Amazon SSM clients implement this trait.