Struct rusoto_ssm::Activation

pub struct Activation {
    pub activation_id: Option<String>,
    pub created_date: Option<f64>,
    pub default_instance_name: Option<String>,
    pub description: Option<String>,
    pub expiration_date: Option<f64>,
    pub expired: Option<bool>,
    pub iam_role: Option<String>,
    pub registration_limit: Option<i64>,
    pub registrations_count: Option<i64>,
}

An activation registers one or more on-premises servers or virtual machines (VMs) with AWS so that you can configure those servers or VMs using Run Command. A server or VM that has been registered with AWS is called a managed instance.

Fields
activation_id: Option<String>

The ID created by Systems Manager when you submitted the activation.

created_date: Option<f64>

The date the activation was created.

default_instance_name: Option<String>

A name for the managed instance when it is created.

description: Option<String>

A user defined description of the activation.

expiration_date: Option<f64>

The date when this activation can no longer be used to register managed instances.

expired: Option<bool>

Whether or not the activation is expired.

iam_role: Option<String>

The Amazon Identity and Access Management (IAM) role to assign to the managed instance.

registration_limit: Option<i64>

The maximum number of managed instances that can be registered using this activation.

registrations_count: Option<i64>

The number of managed instances already registered with this activation.