Struct rusoto_ssm::CommandInvocation

pub struct CommandInvocation {
    pub command_id: Option<String>,
    pub command_plugins: Option<Vec<CommandPlugin>>,
    pub comment: Option<String>,
    pub document_name: Option<String>,
    pub instance_id: Option<String>,
    pub instance_name: Option<String>,
    pub notification_config: Option<NotificationConfig>,
    pub requested_date_time: Option<f64>,
    pub service_role: Option<String>,
    pub standard_error_url: Option<String>,
    pub standard_output_url: Option<String>,
    pub status: Option<String>,
    pub status_details: Option<String>,
    pub trace_output: Option<String>,
}

An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user executes SendCommand against three instances, then a command invocation is created for each requested instance ID. A command invocation returns status and detail information about a command you executed.

Fields
command_id: Option<String>

The command against which this invocation was requested.

command_plugins: Option<Vec<CommandPlugin>> comment: Option<String>

User-specified information about the command, such as a brief description of what the command should do.

document_name: Option<String>

The document name that was requested for execution.

instance_id: Option<String>

The instance ID in which this invocation was requested.

instance_name: Option<String>

The name of the invocation target. For Amazon EC2 instances this is the value for the aws:Name tag. For on-premises instances, this is the name of the instance.

notification_config: Option<NotificationConfig>

Configurations for sending notifications about command status changes on a per instance basis.

requested_date_time: Option<f64>

The time and date the request was sent to this instance.

service_role: Option<String>

The IAM service role that Run Command uses to act on your behalf when sending notifications about command status changes on a per instance basis.

standard_error_url: Option<String>

The URL to the plugin's StdErr file in Amazon S3, if the Amazon S3 bucket was defined for the parent command. For an invocation, StandardErrorUrl is populated if there is just one plugin defined for the command, and the Amazon S3 bucket was defined for the command.

standard_output_url: Option<String>

The URL to the plugin's StdOut file in Amazon S3, if the Amazon S3 bucket was defined for the parent command. For an invocation, StandardOutputUrl is populated if there is just one plugin defined for the command, and the Amazon S3 bucket was defined for the command.

status: Option<String>

Whether or not the invocation succeeded, failed, or is pending.

status_details: Option<String>

A detailed status of the command execution for each invocation (each instance targeted by the command). StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Run Command Status. StatusDetails can be one of the following values:

Pending: The command has not been sent to the instance.

In Progress: The command has been sent to the instance but has not reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command was not delivered to the instance before the delivery timeout expired. Delivery timeouts do not count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution was not complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command was not successful on the instance. For a plugin, this indicates that the result code was not zero. For a command invocation, this indicates that the result code for one or more plugins was not zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

trace_output: Option<String>

Gets the trace output sent by the agent.