Struct rusoto_ssm::CommandPlugin

pub struct CommandPlugin {
    pub name: Option<String>,
    pub output: Option<String>,
    pub output_s3_bucket_name: Option<String>,
    pub output_s3_key_prefix: Option<String>,
    pub output_s3_region: Option<String>,
    pub response_code: Option<i64>,
    pub response_finish_date_time: Option<f64>,
    pub response_start_date_time: Option<f64>,
    pub standard_error_url: Option<String>,
    pub standard_output_url: Option<String>,
    pub status: Option<String>,
    pub status_details: Option<String>,
}

Describes plugin details.

Fields
name: Option<String>

The name of the plugin. Must be one of the following: aws:updateAgent, aws:domainjoin, aws:applications, aws:runPowerShellScript, aws:psmodule, aws:cloudWatch, aws:runShellScript, or aws:updateSSMAgent.

output: Option<String>

Output of the plugin execution.

output_s3_bucket_name: Option<String>

The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

test_folder/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-1234567876543/awsrunShellScript

test_folder is the name of the Amazon S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-1234567876543 is the instance ID;

awsrunShellScript is the name of the plugin.

output_s3_key_prefix: Option<String>

The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

test_folder/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-1234567876543/awsrunShellScript

test_folder is the name of the Amazon S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-1234567876543 is the instance ID;

awsrunShellScript is the name of the plugin.

output_s3_region: Option<String>

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon S3 bucket region.

response_code: Option<i64>

A numeric response code generated after executing the plugin.

response_finish_date_time: Option<f64>

The time the plugin stopped executing. Could stop prematurely if, for example, a cancel command was sent.

response_start_date_time: Option<f64>

The time the plugin started executing.

standard_error_url: Option<String>

The URL for the complete text written by the plugin to stderr. If execution is not yet complete, then this string is empty.

standard_output_url: Option<String>

The URL for the complete text written by the plugin to stdout in Amazon S3. If the Amazon S3 bucket for the command was not specified, then this string is empty.

status: Option<String>

The status of this plugin. You can execute a document with multiple plugins.

status_details: Option<String>

A detailed status of the plugin execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Run Command Status. StatusDetails can be one of the following values:

Pending: The command has not been sent to the instance.

In Progress: The command has been sent to the instance but has not reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command was not delivered to the instance before the delivery timeout expired. Delivery timeouts do not count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution was not complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command was not successful on the instance. For a plugin, this indicates that the result code was not zero. For a command invocation, this indicates that the result code for one or more plugins was not zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist, or it might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit, and they don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.