Struct rusoto_ssm::SsmClient

pub struct SsmClient { /* fields omitted */ }

A client for the Amazon SSM API.

Methods

impl SsmClient

pub fn new(region: Region) -> SsmClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> SsmClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,